### 2019-07-09 #30 @whedon
[PRE REVIEW]: Modia3D: Modeling and Simulation of 3D-Systems in Julia
**Submitting author:** @MartinOtter (<a href="http://orcid.org/0000-0001-6348-9569">Martin Otter</a>)
**Repository:** <a href="https://github.com/ModiaSim/Modia3D.jl" target ="_blank">https://github.com/ModiaSim/Modia3D.jl</a>

Thanks for submitting your paper to JuliaCon @MartinOtter. The JuliaCon editor @matbesancon, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@MartinOtter if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
