### 2019-07-09 #30 @matbesancon
@MartinOtter see the error message, enumitem.sty required
