### 2019-07-10 #30 @MartinOtter
> @MartinOtter see the error message, enumitem.sty required

Added the missing file.
