### 2019-07-10 1bccbfe
Contact with more realistic cushion corrected + additional log output (contact normal/position)
