+++ test/collision/contactForceLaw/contactForceLaw_Billard_BallCushion.jl
@@ -19,9 +19,8 @@ TableX = 2.24
 TableY = 1.12
 TableZ = 0.05
 LyBox = 0.05
-hz = 7*radius/5
+hz  = 7*radius/5
 hzz = hz/3
-
 
 
 @assembly Table(world) begin
@@ -30,11 +29,11 @@ hzz = hz/3
 end
 
 @assembly Cushion(world) begin
-  cushionPart11 = Modia3D.Solid(Modia3D.SolidBox(LyBox, TableY, hz) , "BilliardCushion", vmatTable; contactMaterial = cmatCushion)
-  cushionPart12 = Modia3D.Solid(Modia3D.SolidBox(2.0*LyBox, TableY, hzz) , "BilliardCushion", vmatTable; contactMaterial = cmatCushion)
+  cushionPart11 = Modia3D.Solid(Modia3D.SolidBox(LyBox, TableY, hz) , "BilliardCushion", vmatTable)
+  cushionPart12 = Modia3D.Solid(Modia3D.SolidBox(2.0*LyBox, TableY, hzz,rsmall=0.0) , "BilliardCushion", vmatTable; contactMaterial = cmatCushion)
 
-  cushion11 = Modia3D.Object3D(world, cushionPart11, r=[TableX/2, 0.0, hz/2], fixed=true)
-  cushion12 = Modia3D.Object3D(world, cushionPart12, r=[TableX/2, 0.0, hz + hzz/2], fixed=true)
+  cushion11 = Modia3D.Object3D(world    , cushionPart11, r=[TableX/2, 0.0, hz/2], fixed=true)
+  cushion12 = Modia3D.Object3D(cushion11, cushionPart12, r=[0.0     , 0.0, hz/2+hzz/2], fixed=true)
 end
 
 
@@ -51,9 +50,8 @@ dist = 0.0001
   world = Modia3D.Object3D(visualizeFrame=true)
   table = Table(world)
   cushion = Cushion(world)
-  ballStart = Modia3D.Object3D(world, Modia3D.Solid(Modia3D.SolidSphere(diameter), "BilliardBall", vmatBalls ; contactMaterial = cmatBall), fixed = false, r=[0.0, 0.0, diameter/2], v_start=[6.0, 0.0, 0.0], visualizeFrame=true)
+  ball1 = Modia3D.Object3D(world, Modia3D.Solid(Modia3D.SolidSphere(diameter), "BilliardBall", vmatBalls ; contactMaterial = cmatBall), fixed = false, r=[0.0, 0.0, diameter/2], v_start=[6.0, 0.0, 0.0], visualizeFrame=true)
 end
-
 
 gravField = Modia3D.UniformGravityField(g=9.81, n=[0,0,-1])
 bill = Billiard(sceneOptions=Modia3D.SceneOptions(gravityField=gravField,visualizeFrames=false, defaultFrameLength=0.2,nz_max = 100, enableContactDetection=true, visualizeContactPoints=false, visualizeSupportPoints=false))
@@ -63,15 +61,13 @@ bill = Billiard(sceneOptions=Modia3D.SceneOptions(gravityField=gravField,visuali
 
 model = Modia3D.SimulationModel( bill )
 # ModiaMath.print_ModelVariables(model)
-result = ModiaMath.simulate!(model; stopTime=5.0, tolerance=1e-10,interval=0.001, log=true)
+result = ModiaMath.simulate!(model; stopTime=0.3, tolerance=1e-8,interval=0.001, log=true)
 
 ModiaMath.plot(result, [("ball1.r[1]"),
-                        ("ball1.r[2]"),
                         ("ball1.r[3]"),
                         ("ball1.v[1]"),
                         ("ball1.v[3]"),
[…]
