+++ test/collision/contactForceLaw/contactForceLaw_newtons_cradle.jl
@@ -5,22 +5,36 @@ using Modia3D.StaticArrays
 import Modia3D.ModiaMath
 
 
-vmatGraphics = Modia3D.Material(color="LightBlue" , transparency=0.5)    # material of Graphics
-vmatSolids = Modia3D.Material(color="Red" , transparency=0.5)         # material of solids
+vmatGraphics = Modia3D.Material(color="LightBlue" , transparency=0.1)    # material of Graphics
+vmatSolids = Modia3D.Material(color="Red" , transparency=0.0)         # material of solids
 
 
 cmat = Modia3D.ElasticContactMaterial2("Steel")
 
 Lx = 0.3
-Ly = 4.2
+Ly = 4.1
 Lz = 0.1
 
 
-@assembly NewtonsCradle begin
-  world = Modia3D.Object3D(visualizeFrame=true)
-  box    = Modia3D.Object3D(world, Modia3D.Box(Lx, Ly, Lz; material=vmatGraphics) )
 
-  frame1 = Modia3D.Object3D(world; r=[0.0, -Ly/2, 0.0])
+diameter = 1.0
+lengthWire = 4.0
+
+
+
+@assembly NewtonComparison begin
+  world = Modia3D.Object3D(visualizeFrame=false)
+  cradle = NewtonsCradle(world)
+#  cradle2 = NewtonsCradleSolidWire(world)
+end
+
+
+
+@assembly NewtonsCradle(world) begin
+#   world = Modia3D.Object3D(visualizeFrame=false)
+  box    = Modia3D.Object3D(world, Modia3D.Box(Lx, Ly, Lz; material=vmatGraphics)   ; r=[0.0, 0.0, 0.0] )
+
+  frame1 = Modia3D.Object3D(box; r=[0.0, -Ly/2, 0.0])
 
   frame2 = Modia3D.Object3D(box; r=[0.0, -Ly/4, 0.0])
   frame3 = Modia3D.Object3D(box; r=[0.0, 0.0, 0.0])
@@ -37,12 +51,9 @@ Lz = 0.1
   rev1   = Modia3D.Revolute(frame1, pendulum1.frame1; axis = 1 , phi_start =  -pi/3)
   rev2   = Modia3D.Revolute(frame2, pendulum2.frame1; axis = 1 , phi_start =  -pi/3)
   rev3   = Modia3D.Revolute(frame3, pendulum3.frame1; axis = 1 ) #, phi_start =  -pi/3)
-  rev4   = Modia3D.Revolute(frame4, pendulum4.frame1; axis = 1 , phi_start =  pi/3 ) #
-  rev5   = Modia3D.Revolute(frame5, pendulum5.frame1; axis = 1 , phi_start =  pi/3 ) #
+  rev4   = Modia3D.Revolute(frame4, pendulum4.frame1; axis = 1 )#, phi_start =  pi/3 ) #
+  rev5   = Modia3D.Revolute(frame5, pendulum5.frame1; axis = 1  , phi_start =  pi/3 ) #
 end
-
-diameter = 1.0
-lengthWire = 4.0
 
 @assembly Pendulum() begin
   wire = Modia3D.Object3D( Modia3D.Cylinder(diameter/5, lengthWire; material=vmatGraphics) )
@@ -53,18 +64,56 @@ lengthWire = 4.0
   cyl    = Modia3D.Cylinder(0.15,Lx; material=vmatGraphics)
   cylinder   = Modia3D.Object3D(frame1, cyl; R=ModiaMath.rot2(-pi/2), visualizeFrame=false)
 
[…]
