### 2019-07-10 #30 @Datseris
@matbesancon yes, I accept reviewing this. (but I am letting you know that this will be my first review in JOSS)
