+++ test/collision/contactForceLaw/contactForceLaw_Billard_BallCushion.jl
@@ -6,7 +6,7 @@ import Modia3D.ModiaMath
 
 
 vmatBalls = Modia3D.Material(color="Red" , transparency=0.0)
-vmatTable = Modia3D.Material(color="DarkGreen" , transparency=0.6)
+vmatTable = Modia3D.Material(color="DarkGreen" , transparency=0.4)
 
 cmatTable   = Modia3D.ElasticContactMaterial2("BilliardTable")
 cmatBall    = Modia3D.ElasticContactMaterial2("BilliardBall")
@@ -15,10 +15,10 @@ cmatCushion = Modia3D.ElasticContactMaterial2("BilliardCushion")
 diameter = 0.06
 radius = diameter/2
 
-TableX = 2.24
+TableX = 1.24
 TableY = 1.12
-TableZ = 0.05
-LyBox = 0.05
+TableZ = 0.03
+LyBox = 0.03
 hz  = 7*radius/5
 hzz = hz/3
 
@@ -30,10 +30,18 @@ end
 
 @assembly Cushion(world) begin
   cushionPart11 = Modia3D.Solid(Modia3D.SolidBox(LyBox, TableY, hz) , "BilliardCushion", vmatTable)
-  cushionPart12 = Modia3D.Solid(Modia3D.SolidBox(2.0*LyBox, TableY, hzz,rsmall=0.0) , "BilliardCushion", vmatTable; contactMaterial = cmatCushion)
+  cushionPart12 = Modia3D.Solid(Modia3D.SolidBox(1.5*LyBox, TableY, hzz,rsmall=0.0) , "BilliardCushion", vmatTable; contactMaterial = cmatCushion)
+  cushionPart21 = Modia3D.Solid(Modia3D.SolidBox(TableX, LyBox , hz) , "BilliardCushion", vmatTable)
+  cushionPart22 = Modia3D.Solid(Modia3D.SolidBox(TableX, 1.5*LyBox, hzz,rsmall=0.0) , "BilliardCushion", vmatTable; contactMaterial = cmatCushion)
 
-  cushion11 = Modia3D.Object3D(world    , cushionPart11, r=[TableX/2, 0.0, hz/2], fixed=true)
-  cushion12 = Modia3D.Object3D(cushion11, cushionPart12, r=[0.0     , 0.0, hz/2+hzz/2], fixed=true)
+  cushion11 = Modia3D.Object3D(world    , cushionPart11, r=[TableX/2 - LyBox/2, 0.0, hz/2], fixed=true)
+  cushion12 = Modia3D.Object3D(cushion11, cushionPart12, r=[-LyBox/4     , 0.0, hz/2+hzz/2], fixed=true)
+  cushion31 = Modia3D.Object3D(world    , cushionPart11, r=[-TableX/2 + LyBox/2, 0.0, hz/2], fixed=true)
+  cushion32 = Modia3D.Object3D(cushion31, cushionPart12, r=[LyBox/4     , 0.0, hz/2+hzz/2], fixed=true)
+  cushion21 = Modia3D.Object3D(world    , cushionPart21, r=[0.0, TableY/2 - LyBox/2 , hz/2], fixed=true)
+  cushion22 = Modia3D.Object3D(cushion21, cushionPart22, r=[0.0, -LyBox/4 , hz/2+hzz/2], fixed=true)
+  cushion41 = Modia3D.Object3D(world    , cushionPart21, r=[0.0, -TableY/2 + LyBox/2 , hz/2], fixed=true)
+  cushion42 = Modia3D.Object3D(cushion21, cushionPart22, r=[0.0, LyBox/4 , hz/2+hzz/2], fixed=true)
 end
 
 #=
@@ -50,18 +58,18 @@ dist = 0.0001
[…]
