### 2019-07-13 de1e679
Improved contact pair material (there are still bugs)
