+++ src/Composition/dynamics.jl
@@ -62,11 +62,13 @@ end
     simModel = SimulationModel(assembly::Modia3D.AbstractAssembly;
                                analysis::ModiaMath.AnalysisType=ModiaMath.DynamicAnalysis,
                                startTime = 0.0, stopTime  = 1.0, tolerance = 1e-4,
-                               interval  = (stopTime-startTime)/500.0)
+                               interval  = (stopTime-startTime)/500.0,
+                               maxStepSize = NaN, maxNumberOfSteps=missing)
 
 Generate a `simulationModel` from an `assembly` generated with macro [`Modia3D.@assembly`](@ref)
 and the type of `analysis` to be carried out on the `assembly`.
-Additionally, default `startTime`, `stopTime`, `tolerance`, `interval` for the simulation
+Additionally, default `startTime`, `stopTime`, `tolerance`, `interval`,
+`maxStepSize`, `maxNumberOfSteps`, for the simulation
 engine are defined. These values should be adapted so that assembly-specific, meaningful
 defaults are provided.
 """
@@ -83,6 +85,8 @@ struct SimulationModel <: ModiaMath.AbstractSimulationModel
                             stopTime  = 1.0,
                             tolerance = 1e-4,
                             interval  = (stopTime-startTime)/500.0,
+                            maxStepSize=NaN,
+                            maxNumberOfSteps=missing,
                             hev = 1e-8,
                             scaleConstraintsAtEvents::Bool = true)
       modelName = Modia3D.trailingPartOfName( string( typeof(assembly) ) )
@@ -153,6 +157,8 @@ struct SimulationModel <: ModiaMath.AbstractSimulationModel
                                 defaultStopTime  = stopTime,
                                 defaultTolerance = tolerance,
                                 defaultInterval  = interval,
+                                defaultMaxStepSize = maxStepSize,
+                                defaultMaxNumberOfSteps = maxNumberOfSteps,
                                 getResultNames   = Composition.getResultNames,
                                 getResult        = ModiaMath.Variables.getResult,
                                 storeResult!     = ModiaMath.Variables.storeVariables!,
@@ -252,6 +258,7 @@ end
 
 
 const str_DUMMY = "dummyDistance(nothing,nothing)"
+
 
 function getModelResidues!(m::SimulationModel, time::Float64, _x::Vector{Float64}, _derx::Vector{Float64}, _r::Vector{Float64}, _w::Vector{Float64})
    # println("... time = ", time, ", x = ", _x, ", derx = ", _derx)
@@ -417,7 +424,8 @@ function getModelResidues!(m::SimulationModel, time::Float64, _x::Vector{Float64
       for (pairID, pair) in ch.contactDict
         obj1 = pair.obj1
         obj2 = pair.obj2
-        rContact = (pair.contactPoint1 + pair.contactPoint2)/2.0
+        rContact      = pair.rContact
+        contactNormal = pair.contactNormal
 
         #simh = sim.eventHandler
         #println( "time = ", sim.time, ": ", ModiaMath.instanceName(obj1), " ", ModiaMath.instanceName(obj2),
@@ -429,14 +437,13 @@ function getModelResidues!(m::SimulationModel, time::Float64, _x::Vector{Float64
                 pair.contactPairMaterial = ch.lastContactDict[pairID].contactPairMaterial    # improve later (should avoid to inquire pairID twice)
             else
                 # determine contact pair material
-                delta_dot_init = computeDeltaDotInitial(obj1, obj2, rContact, pair.contactNormal)
[…]
