+++ src/Composition/responseCalculation/elasticCollisionResponse.jl
@@ -1,0 +1,191 @@
+# License for this file: MIT (expat)
+# Copyright 2017-2018, DLR Institute of System Dynamics and Control
+#
+# This file is part of module
+#   Modia3D.Composition (Modia3D/Composition/_module.jl)
+
+"""
+    m = ElasticContactPairResponseMaterial(c_res, d_res, mu_k, mu_r, vsmall, wsmall)
+
+Return an `ElasticContactPairResponseMaterial < AbstractContactPairMaterial` object.
+"""
+mutable struct ElasticContactPairResponseMaterial <: Modia3D.AbstractContactPairMaterial
+    c_res::Float64
+    c_geo::Float64
+    n_geo::Float64
+    d_res::Float64
+    mu_k::Float64
+    mu_r::Float64
+    mu_r_geo::Float64
+    vsmall::Float64
+    wsmall::Float64
+end
+
+
+regularize(absv,vsmall) = absv >= vsmall ? absv : absv*(absv/vsmall)*(1.0 - (absv/vsmall)/3.0) + vsmall/3.0
+
+
+resultantCoefficientOfRestitution(cor, abs_vreln, vsmall; cor_min=0.01) =
+    cor + (cor_min - cor)*exp(log(0.01)*(abs_vreln/vsmall))
+
+
+function resultantDampingCoefficient(cor, abs_vreln, vsmall; cor_min=0.01, d_max=1000.0)
+    @assert(cor >= 0.0 && cor <= 1.0)
+    @assert(abs_vreln >= 0.0)
+    @assert(vsmall  > 0.0)
+    @assert(cor_min > 0.0)
+    @assert(d_max   > 0.0)
+
+    cor_res = resultantCoefficientOfRestitution(cor,abs_vreln,vsmall;cor_min=cor_min)
+    return min(d_max, 8.0*(1.0 - cor_res)/(5*cor_res*regularize(abs_vreln,vsmall)))
+end
+
+
+function elasticContactPairCoefficients(obj1::Object3D, obj2::Object3D)
+  if typeof(obj1.data.geo) == Modia3D.Solids.SolidSphere && typeof(obj2.data.geo) == Modia3D.Solids.SolidSphere
+    r1 = obj1.data.geo.Dx*0.5
+    r2 = obj2.data.geo.Dx*0.5
+    mu_r_geo = r1*r2/(r1 + r2)
+    n_geo = 1.5
+    c_geo = 4/3*sqrt(mu_r_geo)
+  elseif typeof(obj1.data.geo) == Modia3D.Solids.SolidSphere && typeof(obj2.data.geo) != Modia3D.Solids.SolidSphere
+    mu_r_geo = obj1.data.geo.Dx*0.5
+    n_geo = 1.5
+    c_geo = 4/3*sqrt(mu_r_geo)
+  elseif typeof(obj1.data.geo) != Modia3D.Solids.SolidSphere && typeof(obj2.data.geo) == Modia3D.Solids.SolidSphere
+    mu_r_geo = obj2.data.geo.Dx*0.5
+    n_geo = 1.5
+    c_geo = 4/3*sqrt(mu_r_geo)
+  else
+    mu_r_geo = 1.0
+    n_geo = 1.0
+    c_geo = 1.0
+  end
+  return (c_geo, n_geo, mu_r_geo)
+end
+
+
+"""
+   resonseMaterial = contactStart(matPair::Modia3D.ElasticContactPairMaterial,
+                                  obj1,obj2,rContact,contactNormal)
+
+Return a `responseMaterial::ElasticContactPairResponseMaterial` object
+at the start of a collision.
+"""
+function contactStart(matPair::Solids.ElasticContactPairMaterial,
+                      obj1::Object3D,
[…]
