+++ src/Solids/contactPairMaterials.jl
@@ -1,0 +1,147 @@
+# License for this file: MIT (expat)
+# Copyright 2017-2019, DLR Institute of System Dynamics and Control
+#
+# This file is part of module
+#   Modia3D.Solids(Modia3D/Solids/_module.jl)
+#
+
+
+"""
+    material = ElasticContactPairMaterial(;coefficientOfRestitution=0.0, slidingFrictionCoefficient=0.0,
+                                           rotationalResistanceCoefficient=0.0, vsmall=0.01, wsmall=0.01)
+
+Generates an `ElasticContactPairMaterial` object by providing the material properties of
+two objects that are in contact to each other.
+
+# Keyword Arguments
+- coefficientOfRestitution: Coefficient of restitution between two objects (=0: inelastic ... =1: fully elastic).
+- slidingFrictionCoefficient: Kinetic/sliding friction force coefficent between two objects (`>= 0.0`).
+- rotationalResistanceCoefficient: Rotational resistance torque coefficient between two objects (`>= 0.0`).
+  Its effect is that the contact torque is computed to reduce the relative angular velocity between two objects.
+  For a ball, `rotationalResistanceCoefficient` is the (standard) rolling resistance coefficient.
+- `vsmall` in [m/s]: Used for regularization when computing the unit vector in direction of
+                     the relative tangential velocity to avoid a division by zero.
+- `wsmall` in [rad/s]: Used for regularization when computing the unit vector in direction
+                       of the relative angular velocity to avoid a division by zero.
+
+# Example
+```julia
+import Modia3D
+mat = Modia3D.ElasticContactPairMaterial(coefficientOfRestitution=0.5)
+```
+"""
+mutable struct ElasticContactPairMaterial <: Modia3D.AbstractContactPairMaterial
+    coefficientOfRestitution::Float64
+    slidingFrictionCoefficient::Float64
+    rotationalResistanceCoefficient::Float64
+    vsmall::Float64
+    wsmall::Float64
+
+    function ElasticContactPairMaterial(; coefficientOfRestitution=0.0,
+                                          slidingFrictionCoefficient=0.0,
+                                          rotationalResistanceCoefficient=0.0,
+                                          vsmall=0.01,
+                                          wsmall=0.01)
+        @assert(coefficientOfRestitution >= 0.0)
+        @assert(coefficientOfRestitution <= 1.0)
+        @assert(slidingFrictionCoefficient >= 0.0)
+        @assert(rotationalResistanceCoefficient >= 0.0)
+        new(coefficientOfRestitution, slidingFrictionCoefficient, rotationalResistanceCoefficient,
+            vsmall, wsmall)
+    end
+end
+
+
+"""
+    material = combineContactPairMaterials(material1, material2)
+
+Return a new `ElasticContactPairMaterial` object from the mean values of the ElasticContactPairMaterial
+objects `material1` and `material2`.
+"""
+function combineContactPairMaterials(mat1::ElasticContactPairMaterial, mat2::ElasticContactPairMaterial)::ElasticContactPairMaterial
+    mat = ElasticContactPairMaterial()
+    mat.coefficientOfRestitution        = (mat1.coefficientOfRestitution        + mat2.coefficientOfRestitution)/2
+    mat.slidingFrictionCoefficient      = (mat1.slidingFrictionCoefficient      + mat2.slidingFrictionCoefficient)/2
+    mat.rotationalResistanceCoefficient = (mat1.rotationalResistanceCoefficient + mat2.rotationalResistanceCoefficient)/2
+    mat.vsmall = (mat1.vsmall + mat2.vsmall)/2
[…]
