### 2019-07-13 3c2af19
Debug print included
