+++ src/Composition/responseCalculation/elasticCollisionResponse.jl
@@ -130,7 +130,7 @@ at time `time`.
 """
 function responseCalculation(material::ElasticContactPairResponseMaterial, obj1::Object3D, obj2::Object3D,
                              rContact::ModiaMath.Vector3D, e_n::ModiaMath.Vector3D,
-                             s::Float64, time::Float64)::Tuple{ModiaMath.Vector3D,ModiaMath.Vector3D,ModiaMath.Vector3D,ModiaMath.Vector3D}
+                             s::Float64, time::Float64, file)::Tuple{ModiaMath.Vector3D,ModiaMath.Vector3D,ModiaMath.Vector3D,ModiaMath.Vector3D}
     # Material data
     c_res    = material.c_res
     c_geo    = material.c_geo
@@ -182,10 +182,24 @@ function responseCalculation(material::ElasticContactPairResponseMaterial, obj1:
     t1  = cross(r_rel1,f1) + tau
     t2  = cross(r_rel2,f2) - tau
 
-    if time > 0.022
-        #println("obj1= \"", ModiaMath.instanceName(obj1), "\" obj2 = ", ModiaMath.instanceName(obj2), " fn = ", fn, " en = ", e_n, " delta_dot ", delta_dot, " time = ", time, " mu_k = ", mu_k, " mu_r = ", mu_r)
-        println(" fn = ", fn, " ft = ", ft, " tau = ", tau, " e_n = ", e_n, " delta_dot ", delta_dot, " time = ", time, " mu_k = ", mu_k, " mu_r = ", mu_r)
-     end
+  println(file,"... time = ", time,
+          " fn = ", fn, 
+          " ft = ", ft, 
+          " tau = ", tau, 
+          " e_n = ", e_n, 
+          " delta_dot ", delta_dot, 
+          " mu_k = ", mu_k, 
+          " mu_r = ", mu_r,
+          " c_geo = ", c_geo,
+          " n_geo = ", n_geo,
+          " mu_r_geo = ", mu_r_geo,
+          " f1 = ", f1,
+          " t1 = ", t1)
+
+    #if time > 0.022
+    #    #println("obj1= \"", ModiaMath.instanceName(obj1), "\" obj2 = ", ModiaMath.instanceName(obj2), " fn = ", fn, " en = ", e_n, " delta_dot ", delta_dot, " time = ", time, " mu_k = ", mu_k, " mu_r = ", mu_r)
+    #    println(" fn = ", fn, " ft = ", ft, " tau = ", tau, " e_n = ", e_n, " delta_dot ", delta_dot, " time = ", time, " mu_k = ", mu_k, " mu_r = ", mu_r)
+    # end
 
     return (f1,f2,t1,t2)
 end
