### 2019-07-14 5de20a4
Cleaned-up versions of collision examples added
