+++ src/Solids/solid.jl
@@ -57,12 +57,12 @@ struct Solid <: Modia3D.AbstractObject3Ddata
    geo::Union{Modia3D.AbstractSolidGeometry,NOTHING}
    massProperties::Union{Solids.MassProperties,NOTHING}
    material::Union{Graphics.Material,NOTHING}
-   contactMaterial::Union{Modia3D.AbstractContactMaterial,NOTHING}
+   contactMaterial::Union{String,Modia3D.AbstractContactMaterial,NOTHING}
 
    function Solid(geo::Union{Modia3D.AbstractSolidGeometry,NOTHING},
                   massProperties::Union{Solids.MassProperties,Number,AbstractString,Solids.SolidMaterial,NOTHING} = nothing,
                   material::Union{Graphics.Material,NOTHING} = Graphics.Material();
-                  contactMaterial::Union{Modia3D.AbstractContactMaterial,AbstractString,NOTHING} = nothing)::Solid
+                  contactMaterial="")::Solid
       if typeof(massProperties) == NOTHING
          solid = new(geo, nothing, material, contactMaterial)
       else
@@ -76,7 +76,7 @@ struct Solid <: Modia3D.AbstractObject3Ddata
    function Solid(geo::Vector{Modia3D.Solids.SolidFileMesh},
                   massProperties::Union{Solids.MassProperties,Number,AbstractString,Solids.SolidMaterial,NOTHING},
                   material::Union{Graphics.Material,NOTHING} = Graphics.Material();
-                  contactMaterial::Union{Modia3D.AbstractContactMaterial,AbstractString,NOTHING} = nothing)::Solid
+                  contactMaterial="")::Solid
      solids = Vector{Solid}()
      for geoParts in geo
        push!(solids, Solid(geoParts, massProperties, material; contactMaterial=contactMaterial))

[…]
