### 2019-07-14 efcdf56
Removed REQUIRE file and updated Project.toml, Manifest.toml files
