### 2019-07-15 e8ba3bc
canCollide() adapted to improved collision material; new example Simulate_Robot started
