+++ objects/robot/license.txt
@@ -1,0 +1,81 @@
+The files in this directory had been extracted from the blender file
+   http://www.blendswap.com/blends/view/67329
+The license of this Blender file is given below
+
+--------------------------------------------------------------------------------
+Thank you for downloading "Rigged Robotic Arm" by raovq
+Released under
+Creative Commons Zero (Public Domain)
+Downloaded from http://www.blendswap.com/blends/view/67329
+
+The following is important licensing information, please keep this file for
+archival and future reference when working with the file. It's important that
+you know, understand and follow any requirements stated in this file regarding
+the usage of the materials included.
+
+Some parts of this file are marked with [#], the corresponding numbered note is
+at the end of the file.
+
+--------------------------------------------------------------------------------
+################################################################################
+
+VERY IMPORTANT LICENSE INFORMATION:
+
+This blend has been released under
+Creative Commons Zero (Public Domain)
+
+This means that you can use it for any purpose you see fit, even commercially,
+as long as you respect these requirements:
+
+	--There are no requirements for CC-0 licensed blends.
+################################################################################
+--------------------------------------------------------------------------------
+
+ABOUT THE BLEND:
+
+Ready for Blender 2.65
+Published on: 2013-03-27 21:01:57
+
+--------------------------------------------------------------------------------
+
+HELP US MODERATE THIS BLEND:
+If you find anomalies in this blend or any of the contained files such as:
+
+	- Missing, unneeded or corrupted files.
+	- Inaccurate/mismatching preview image on the site.
+	- Illegal distribution of third party files.
+	- Ripping from a game or other 3D repository.
+	- Uncredited or incorrect use of other CC licensed works.
+	- Some other troubling issues[1].
+
+Please submit a report from  http://www.blendswap.com/blends/view/67329
+by pressing the red button with the flag, including links and details that serve
+as evidence and can help us to solve any conflicts or issues derived from the
+contents of this file.
+
+--------------------------------------------------------------------------------
+
+Thank you for using Blend Swap!
+
+Register to get tons of more blends! http://www.blendswap.com/register
+Share your own blends with the world from http://www.blendswap.com/blends/add
+
+Get answers to your questions and doubts: http://www.blendswap.com/page/faq
+If the site doesn't work for you try here: http://www.blendswap.com/page/issues
+For questions please contact us: http://www.blendswap.com/contact
+Check out our Terms Of Use: http://www.blendswap.com/tou
+Report website bugs: http://www.blendswap.com/bugs
+
+Consider getting an associate Membership to get some neat features and
+enhancements.
+
+--------------------------------------------------------------------------------
+
+NOTES:
+
+[1] Please make sure your problem is not derived from a setting in Blender (like
+hidden layers or objects) before submitting a report under the "Other" category.
+Issues arising from this type of problem will be ignored.
+
+--------------------------------------------------------------------------------
+######################    END OF BLEND SWAP LICENSE.txt    #####################
