### 2019-07-15 c94b4f9
KUKA YouBot added as example (initial version; only kinematic movement)
