+++ examples/collisions/Simulate_YouBot_withTable.jl
@@ -1,0 +1,164 @@
+module Simulate_YouBot
+
+using  Modia3D
+import Modia3D.ModiaMath
+using  Modia3D.ModiaMath.Unitful
+
+vmat1 = Modia3D.Material(color="LightBlue" , transparency=0.3)    # material of SolidFileMesh
+vmat2 = Modia3D.Material(color="Red" , transparency=0.0) # material of solidThing
+
+
+base_frame_obj           = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/base_frame.obj")
+plate_obj                = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/plate.obj")
+front_right_wheel_obj    = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/front-right_wheel.obj")
+front_left_wheel_obj     = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/front-left_wheel.obj")
+back_right_wheel_obj     = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/back-right_wheel.obj")
+back_left_wheel_obj      = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/back-left_wheel.obj")
+
+arm_base_frame_obj       = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/arm_base_frame.obj")
+arm_joint_1_obj          = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/arm_joint_1.obj")
+arm_joint_2_obj          = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/arm_joint_2.obj")
+arm_joint_3_obj          = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/arm_joint_3.obj")
+arm_joint_4_obj          = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/arm_joint_4.obj")
+arm_joint_5_obj          = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/arm_joint_5.obj")
+gripper_base_frame_obj   = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/gripper_base_frame.obj")
+gripper_left_finger_obj  = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/gripper_left_finger.obj")
+gripper_right_finger_obj = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/gripper_right_finger.obj")
+
+
+# Damper
+@forceElement Damper(; d=1.0) begin
+    w   = ModiaMath.RealScalar("w",   causality=ModiaMath.Input,  numericType=ModiaMath.WR)
+    tau = ModiaMath.RealScalar("tau", causality=ModiaMath.Output, numericType=ModiaMath.WR)
+end
+function Modia3D.computeTorque(damper::Damper, sim::ModiaMath.SimulationState)
+    damper.tau.value = -damper.d*damper.w.value
+end
+
+
+# Controller
+@forceElement P_PI_Controller(; k1=1.0, k2=1.0, T2=0.01, freqHz=0.5, A=1.0) begin
+    PI_x    = ModiaMath.RealScalar("PI_x"   , start=0.0, info="State of PI controller", numericType=ModiaMath.XD_EXP)
+    PI_derx = ModiaMath.RealScalar("PI_derx", start=0.0, info="= der(PI_x)"           , numericType=ModiaMath.DER_XD_EXP, integral=PI_x)
[…]
