### 2019-07-16 52c78ba
obj-files corrected and YouBot model simplified
