+++ examples/collisions/Simulate_YouBot.jl
@@ -4,7 +4,7 @@ using  Modia3D
 import Modia3D.ModiaMath
 using  Modia3D.ModiaMath.Unitful
 
-vmat1 = Modia3D.Material(color="LightBlue" , transparency=0.4)    # material of SolidFileMesh
+vmat1 = Modia3D.Material(color="LightBlue" , transparency=0.0)    # material of SolidFileMesh
 
 base_frame_obj           = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/base_frame.obj")
 plate_obj                = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/plate.obj")
@@ -60,13 +60,11 @@ function Modia3D.computeSignal(signal::Sine, sim::ModiaMath.SimulationState)
 end
 
 
-@assembly Link(frame_a; axis=3, R_a_rev=ModiaMath.NullRotation, r_rev_b=[0.0, 0.0, 0.0],
-                r_rev_CM=[0.0, 0.0, 0.0], m=0.0, Ixx=0.0, Iyy=0.0, Izz=0.0, fileMesh="",
+@assembly Link(frame_a; axis=3, R_a_rev=ModiaMath.NullRotation, r_rev_b=[0.0, 0.0, 0.0], m=0.0, fileMesh="",
                 d=0.0, k1=10.0, k2=10.0, T2=0.01, freqHz=0.1, A=1.0) begin
     rev_a   = Object3D(frame_a, R=R_a_rev)
-    rev_b   = Object3D(Solid(nothing,MassProperties(m=m, rCM=r_rev_CM, Ixx=Ixx, Iyy=Iyy, Izz=Izz),vmat1), visualizeFrame=true)
+    rev_b   = Object3D(Solid(SolidFileMesh(fileMesh),m,vmat1), visualizeFrame=false)
     rev     = Revolute(rev_a, rev_b)
-    frame_vis = Object3D(rev_b, FileMesh(fileMesh), R=ModiaMath.rot1(90u"°"))
     frame_b = Object3D(rev_b, r=r_rev_b)
 
     # Kinematic movement of joint
@@ -89,61 +87,38 @@ end
 
 @assembly Gripper(frame_a) begin
     frame1                   = Object3D(frame_a, R=ModiaMath.rot3(-180u"°"))
-    gripper_base_frame       = Object3D(frame1,Solid(nothing,MassProperties(m=0.199, Ixx=0.0002324, Iyy=0.0003629, Izz=0.0002067),nothing), visualizeFrame=true)
-    gripper_base_frame_vis   = Object3D(frame1, FileMesh(gripper_base_frame_obj), R=ModiaMath.rot1(90u"°"))
-
-    gripper_left_finger      = Object3D(gripper_base_frame, Solid(nothing,MassProperties(m=0.010),nothing), r=[0, 0.0082,0])
-    gripper_left_finger_vis  = Object3D(gripper_left_finger, FileMesh(gripper_left_finger_obj), R=ModiaMath.rot1(90u"°"))
-
-    gripper_right_finger     = Object3D(gripper_base_frame, Solid(nothing,MassProperties(m=0.010),nothing), r=[0,-0.0082,0])
-    gripper_right_finger_vis = Object3D(gripper_right_finger, FileMesh(gripper_right_finger_obj), R=ModiaMath.rot1(90u"°"))
+    gripper_base_frame       = Object3D(frame1,Solid(SolidFileMesh(gripper_base_frame_obj),0.199,vmat1), visualizeFrame=true)
+    gripper_left_finger      = Object3D(gripper_base_frame, Solid(SolidFileMesh(gripper_left_finger_obj),0.010,vmat1), r=[0, 0.0082,0])
[…]
