### 2019-07-17 569c3ab
Simulate_YouBot example: minor cleanup
