+++ examples/collisions/Simulate_YouBot.jl
@@ -4,7 +4,7 @@ using  Modia3D
 import Modia3D.ModiaMath
 using  Modia3D.ModiaMath.Unitful
 
-vmat1 = Modia3D.Material(color="LightBlue" , transparency=0.0)    # material of SolidFileMesh
+vmat1 = Modia3D.Material(color="LightBlue" , transparency=0.8)    # material of SolidFileMesh
 
 base_frame_obj           = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/base_frame.obj")
 plate_obj                = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/plate.obj")
@@ -35,7 +35,7 @@ end
 
 
 # Controller
-@forceElement P_PI_Controller(; k1=1.0, k2=1.0, T2=0.01, freqHz=0.5, A=1.0) begin
+@forceElement P_PI_Controller(; k1=1.0, k2=10.0, T2=0.01, freqHz=0.05, A=1.0) begin
     PI_x    = ModiaMath.RealScalar("PI_x"   , start=0.0, info="State of PI controller", numericType=ModiaMath.XD_EXP)
     PI_derx = ModiaMath.RealScalar("PI_derx", start=0.0, info="= der(PI_x)"           , numericType=ModiaMath.DER_XD_EXP, integral=PI_x)
     sine_y  = ModiaMath.RealScalar("sine_y" , start=0.0, info="= sin(2*pi*f*time)"    , numericType=ModiaMath.WR)
@@ -52,32 +52,36 @@ function Modia3D.computeTorque(c::P_PI_Controller, sim::ModiaMath.SimulationStat
 end
 
 
-@signal Sine(;A=1.5, freqHz = 0.5) begin
+@signal Sine(;A=1.5, freqHz = 0.5, offset=0.0) begin
    y = ModiaMath.RealScalar(causality=ModiaMath.Output, numericType=ModiaMath.WR)
 end
 function Modia3D.computeSignal(signal::Sine, sim::ModiaMath.SimulationState)
-    signal.y.value  = signal.A*sin(2*pi*signal.freqHz*sim.time)
+    signal.y.value  = signal.offset + signal.A*sin(2*pi*signal.freqHz*sim.time)
 end
 
 
-@assembly Link(frame_a; axis=3, R_a_rev=ModiaMath.NullRotation, r_rev_b=[0.0, 0.0, 0.0], m=0.0, fileMesh="",
-                d=0.0, k1=10.0, k2=10.0, T2=0.01, freqHz=0.1, A=1.0) begin
-    rev_a   = Object3D(frame_a, R=R_a_rev)
-    rev_b   = Object3D(Solid(SolidFileMesh(fileMesh),m,vmat1), visualizeFrame=false)
-    rev     = Revolute(rev_a, rev_b)
-    frame_b = Object3D(rev_b, r=r_rev_b)
+@assembly Link(frame_a; axis=3, fileMesh="", m=0.0, r_rev_b=[0.0, 0.0, 0.0], R_a_rev=ModiaMath.NullRotation,
+               d=0.0, k1=1.0, k2=10.0, T2=0.01, freqHz=0.1, A=0.1, offset=0.0) begin
+    rev_a   = Object3D(frame_a, R=R_a_rev, visualizeFrame=false)
+    rev_b   = Object3D(Solid(SolidFileMesh(fileMesh),m,vmat1))
+    rev     = Revolute(rev_a, rev_b, phi_start=offset)
+    frame_b = Object3D(rev_b, r=r_rev_b, visualizeFrame=false)
 
+#=
     # Kinematic movement of joint
-    sig         = Sine(A=A,freqHz=freqHz)
+    sig         = Sine(A=A,freqHz=freqHz,offset=offset)
[…]
