### 2019-07-17 3cc4f88
YouBot robot examples: controllers added + table
