+++ examples/collisions/Simulate_YouBot.jl
@@ -4,7 +4,7 @@ using  Modia3D
 import Modia3D.ModiaMath
 using  Modia3D.ModiaMath.Unitful
 
-vmat1 = Modia3D.Material(color="LightBlue" , transparency=0.8)    # material of SolidFileMesh
+vmat1 = Modia3D.Material(color="LightBlue" , transparency=0.0)    # material of SolidFileMesh
 
 base_frame_obj           = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/base_frame.obj")
 plate_obj                = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/plate.obj")
@@ -23,6 +23,30 @@ gripper_base_frame_obj   = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/gri
 gripper_left_finger_obj  = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/gripper_left_finger.obj")
 gripper_right_finger_obj = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/gripper_right_finger.obj")
 
+# Drive train data
+effectiveInertia(J, gearRatio) = J*gearRatio^2
+
+const motorInertia1   = 0.0000135 + 0.000000409
+const gearboxRatio1   = 156.0
+const J1              = effectiveInertia(motorInertia1, gearboxRatio1)
+
+const motorInertia2   = 0.0000135 + 0.000000409
+const gearboxRatio2   = 156.0
+const J2              = effectiveInertia(motorInertia2, gearboxRatio2)
+
+const motorInertia3   = 0.0000135 + 0.000000071
+const gearboxRatio3   = 100.0
+const J3              = effectiveInertia(motorInertia3, gearboxRatio3)
+
+const motorInertia4   = 0.00000925 + 0.000000071
+const gearboxRatio4   = 71.0
+const J4              = effectiveInertia(motorInertia4, gearboxRatio4)
+
+const motorInertia5   = 0.0000035 + 0.000000071
+const gearboxRatio5   = 71.0
+const J5              = effectiveInertia(motorInertia5, gearboxRatio5)
+
+
 
 # Damper
 @forceElement Damper(; d=1.0) begin
@@ -35,7 +59,7 @@ end
 
 
 # Controller
-@forceElement P_PI_Controller(; k1=1.0, k2=10.0, T2=0.01, freqHz=0.05, A=1.0) begin
+@forceElement P_PI_Controller(; k1=1.0, k2=10.0, T2=0.01, freqHz=0.05, A=1.0, gearRatio=1.0) begin
     PI_x    = ModiaMath.RealScalar("PI_x"   , start=0.0, info="State of PI controller", numericType=ModiaMath.XD_EXP)
     PI_derx = ModiaMath.RealScalar("PI_derx", start=0.0, info="= der(PI_x)"           , numericType=ModiaMath.DER_XD_EXP, integral=PI_x)
     sine_y  = ModiaMath.RealScalar("sine_y" , start=0.0, info="= sin(2*pi*f*time)"    , numericType=ModiaMath.WR)
@@ -44,11 +68,12 @@ end
     tau = ModiaMath.RealScalar("tau", causality=ModiaMath.Output, numericType=ModiaMath.WR)
 end
 function Modia3D.computeTorque(c::P_PI_Controller, sim::ModiaMath.SimulationState)
+    gearRatio       = c.gearRatio
     c.sine_y.value  = c.A*sin(2*pi*c.freqHz*sim.time)
[…]
