### 2019-07-17 6f48448
Test YouBot mit table removed
