### 2019-07-17 8715592
Bug with scene.collide temporarily fixed (should be made better) + first version of path planning
