### 2019-07-17 4bf0e84
Further fix of bug with scene.collision
