+++ src/Basics/pathPlanning.jl
@@ -1,0 +1,236 @@
+# License for this file: MIT (expat)
+# Copyright 2017-2018, DLR Institute of System Dynamics and Control
+#
+# This file is part of module
+#   Modia3D.Basics (Modia3D/Basics/_module.jl)
+#
+
+
+"""
+    path = PTP_path(names;
+                    positions = [zeros(size(names,1))'; ones(size(names,1))'],
+                    startTime=0.0,
+                    v_max = ones(size(names,1)),
+                    a_max = ones(size(names,2))))
+
+Generate a new path object.
+"""
+mutable struct PTP_path
+   names::Vector{AbstractString}
+   startTime::Float64
+   v_max::Vector{Float64}
+   a_max::Vector{Float64}
+   positions::Matrix{Float64}
+
+   delta::Matrix{Float64}
+   hasPath::Vector{Bool}
+   sd_max::Vector{Float64}
+   sdd_max::Vector{Float64}
+   Ta1::Vector{Float64}
+   Ta2::Vector{Float64}
+   noWphase::Vector{Bool}
+   Tv::Vector{Float64}
+   Te::Vector{Float64}
+   Ta1s::Vector{Float64}
+   Ta2s::Vector{Float64}
+   Tvs::Vector{Float64}
+   Tes::Vector{Float64}
+   sd_max2::Vector{Float64}
+   s1::Vector{Float64}
+   s2::Vector{Float64}
+   s3::Vector{Float64}
+   Tend::Float64
+
+   function PTP_path(names::AbstractVector;
+                     positions::Matrix{Float64} = [zeros(size(names,1))'; ones(size(names,1))'],
+                     startTime=0.0,
+                     v_max = ones(size(names,1)),
+                     a_max = ones(size(names,2)))
+
+        #@assert(size(positions,1) > 1)
+        #@assert(size(positions,2) == Base.length(names))
+        #@assert(Base.length(v_max) == Base.length(names))
+        #@assert(Base.length(a_max) == Base.length(names))
+        np    = Base.length(names)
+        npath = size(positions,1) - 1   # number of path points
+
+        for i in eachindex(v_max)
+            @assert(v_max[i] > 0.0)
+            @assert(a_max[i] > 0.0)
+        end
+
+        delta    = zeros(npath, np)
+        hasPath  = fill(false,npath)
+        sd_max   = zeros(npath)
+        sdd_max  = zeros(npath)
+        Ta1      = zeros(npath)
+        Ta2      = zeros(npath)
+        noWphase = fill(false,npath)
+        Tv       = zeros(npath)
+        Te       = zeros(npath)
+        Ta1s     = zeros(npath)
+        Ta2s     = zeros(npath)
+        Tvs      = zeros(npath)
+        Tes      = zeros(npath)
+        sd_max2  = zeros(npath)
+        s1       = zeros(npath)
+        s2       = zeros(npath)
+        s3       = zeros(npath)
+        aux1     = zeros(np)
+        aux2     = zeros(np)
+
+        for i in 1:npath
+            delta[i,:] = positions[i+1,:] - positions[i,:]
+
+            for j in 1:np
+                aux1[j] = delta[i,j]/v_max[j]
+                aux2[j] = delta[i,j]/a_max[j]
+            end
+
+            sd_max_inv  = maximum(abs.(aux1))
+            sdd_max_inv = maximum(abs.(aux2))
+
+            small = 1000*eps()
+            if sd_max_inv <= small || sdd_max_inv <= small
+                hasPath[i] = false
+            else
+                hasPath[i] = true
+                sd_max[i]  = 1/sd_max_inv
[…]
