+++ examples/collisions/Simulate_YouBot.jl
@@ -27,31 +27,32 @@ gripper_right_finger_obj = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/gri
 effectiveInertia(J, gearRatio) = J*gearRatio^2
 
 const motorInertia1 = 0.0000135 + 0.000000409
-const gearboxRatio1 = 156.0
-const J1            = effectiveInertia(motorInertia1, gearboxRatio1)
+const gearRatio1    = 156.0
+const J1            = effectiveInertia(motorInertia1, gearRatio1)
 
 const motorInertia2 = 0.0000135 + 0.000000409
-const gearboxRatio2 = 156.0
-const J2            = effectiveInertia(motorInertia2, gearboxRatio2)
+const gearRatio2    = 156.0
+const J2            = effectiveInertia(motorInertia2, gearRatio2)
 
 const motorInertia3 = 0.0000135 + 0.000000071
-const gearboxRatio3 = 100.0
-const J3            = effectiveInertia(motorInertia3, gearboxRatio3)
+const gearRatio3    = 100.0
+const J3            = effectiveInertia(motorInertia3, gearRatio3)
 
 const motorInertia4 = 0.00000925 + 0.000000071
-const gearboxRatio4 = 71.0
-const J4            = effectiveInertia(motorInertia4, gearboxRatio4)
+const gearRatio4    = 71.0
+const J4            = effectiveInertia(motorInertia4, gearRatio4)
 
 const motorInertia5 = 0.0000035 + 0.000000071
-const gearboxRatio5 = 71.0
-const J5            = effectiveInertia(motorInertia5, gearboxRatio5)
+const gearRatio5    = 71.0
+const J5            = effectiveInertia(motorInertia5, gearRatio5)
 
 
-println("\nJ1 = $J1")
-println("J2 = $J2")
-println("J3 = $J3")
-println("J4 = $J4")
-println("J5 = $J5\n")
+println("")
+println("J1 = $J1, gearRatio1 = $gearRatio1")
+println("J2 = $J2, gearRatio2 = $gearRatio2")
+println("J3 = $J3, gearRatio3 = $gearRatio3")
+println("J4 = $J4, gearRatio4 = $gearRatio4")
+println("J5 = $J5, gearRatio5 = $gearRatio5\n")
 
 
 # Damper
@@ -65,28 +66,40 @@ end
 
 
 # Controller
-@forceElement P_PI_Controller(; k1=1.0, k2=10.0, T2=0.01, freqHz=0.05, A=1.0, gearRatio=1.0) begin
+const ptp_path = PTP_path(["angle1", "angle2", "angle3", "angle4", "angle5"],
+                          positions = [0.0 0.0 1.5 0.0 0.0;
+                                       3.0 1.5 0.0 1.3 1.4;
+                                       1.5 0.0 0.5 0.5 0.5;
+                                       0.0 0.0 1.5 0.0 0.0],
+                          startTime = 0.0,
+                          v_max = [2.68512, 2.68512, 4.8879, 5.8997, 5.8997],
+                          a_max = 1.5*ones(5))
+
+plotPath(ptp_path)
+
+@forceElement P_PI_Controller(; k1=1.0, k2=10.0, T2=0.01, gearRatio=1.0, phi_ref_name="UnknownName") begin
+    phi_ref_index = getIndex(ptp_path, phi_ref_name)
[…]
