### 2019-07-18 3e23acc
YouBotRobot with moving gripper (new prismatic joint with damper + P-PI controller)
