### 2019-07-18 9a26e6c
Added test examples (box on table) + improved YouBot example + new SceneOption: elasticContactReductionFactor
