+++ examples/collisions/Simulate_YouBot.jl
@@ -5,6 +5,8 @@ import Modia3D.ModiaMath
 using  Modia3D.ModiaMath.Unitful
 
 vmat1 = Modia3D.Material(color="LightBlue" , transparency=0.0)    # material of SolidFileMesh
+vmat2 = Modia3D.Material(color="Grey"      , transparency=0.3)    # material of work piece
+vmat3 = Modia3D.Material(color="LightBlue" , transparency=0.3)    # material of table
 
 base_frame_obj           = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/base_frame.obj")
 plate_obj                = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/plate.obj")
@@ -66,16 +68,30 @@ end
 
 
 # Controller
-const ptp_path = PTP_path(["angle1", "angle2", "angle3", "angle4", "angle5", "gripper"],
-                          positions = [0.0 0.0 1.5 0.0 0.0 0.0;
-                                       3.0 1.5 0.0 1.3 1.4 0.0;
-                                       1.5 0.0 0.5 0.5 0.5 0.0;
-                                       0.0 0.0 1.5 0.0 0.0 0.0;
-                                       0.0 0.0 1.5 0.0 0.0 0.04;
-                                       0.0 0.0 1.5 0.0 0.0 0.0;],
-                          startTime = 0.0,
-                          v_max = [2.68512, 2.68512, 4.8879, 5.8997, 5.8997, 1.0],
-                          a_max = [1.5, 1.5, 1.5, 1.5, 1.5, 0.5])
+const ptp_path1 = PTP_path(["angle1", "angle2", "angle3", "angle4", "angle5", "gripper"],
+                           positions = [0.0  0.0  pi/2 0.0 0.0 0.0;
+                                        pi   pi/2 0.0  1.3 1.4 0.0;
+                                        pi/2 0.0  0.5  0.5 0.5 0.0;
+                                        0.0  0.0  pi/2 0.0 0.0 0.0;
+                                        0.0  0.0  pi/2 0.0 0.0 0.04;
+                                        0.0  0.0  pi/2 0.0 0.0 0.041;],
+                           startTime = 0.0,
+                           v_max = [2.68512, 2.68512, 4.8879, 5.8997, 5.8997, 1.0],
+                           a_max = [1.5, 1.5, 1.5, 1.5, 1.5, 0.5])
+
+const ptp_path2 = PTP_path(["angle1", "angle2", "angle3", "angle4", "angle5", "gripper"],
+                           positions = [0.0  0.0 0.0  0.0 0.0 0.0;
+                                        0.0  0.0 0.0  0.0 0.0 0.04;
+                                        0.0  0.0 pi/2 0.0 0.0 0.04;
+                                        0.0  0.0 pi/2 0.0 0.0 0.03807;    # no contact 0.381; 0.38068 with contactReductionFactor=1e-4 fails; 0.3807: sliding contact (box remains on the table)
+                                        0.0 -pi/4 2.5*pi/4 0.0 0.0 0.03807;
+                                        0.0  0.0 0.0  0.0 0.0 0.03807],
+                           startTime = 0.0,
+                           v_max = [2.68512, 2.68512, 4.8879, 5.8997, 5.8997, 1.0],
[…]
