### 2019-07-21 d77c6aa
Bug in contact handler corrected + YouBot example improved (impact of tool tip with sphere)
