+++ examples/collisions/Simulate_YouBot.jl
@@ -7,6 +7,7 @@ using  Modia3D.ModiaMath.Unitful
 vmat1 = Modia3D.Material(color="LightBlue" , transparency=0.0)    # material of SolidFileMesh
 vmat2 = Modia3D.Material(color="Grey"      , transparency=0.3)    # material of work piece
 vmat3 = Modia3D.Material(color="LightBlue" , transparency=0.3)    # material of table
+vmat4 = Modia3D.Material(color="Green"     , transparency=0.0)    # material of ground
 
 base_frame_obj           = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/base_frame.obj")
 plate_obj                = joinpath(Modia3D.path, "objects/robot_KUKA_YouBot/plate.obj")
@@ -90,7 +91,14 @@ const ptp_path2 = PTP_path(["angle1", "angle2", "angle3", "angle4", "angle5", "g
                            v_max = [2.68512, 2.68512, 4.8879, 5.8997, 5.8997, 1.0],
                            a_max = [1.5, 1.5, 1.5, 1.5, 1.5, 0.5])
 
-const ptp_path = ptp_path2
+const ptp_path3 = PTP_path(["angle1", "angle2", "angle3", "angle4", "angle5", "gripper"],
+                           positions = [0.0  0.0 pi/2     0.0 0.0 0.0;
+                                        0.0  0.3 pi/2-0.3 0.0 0.0 0.0;
+                                        0.0  0.0 0.0      0.0 0.0 0.0],
+                           startTime = 0.0,
+                           v_max = [2.68512, 2.68512, 4.8879, 5.8997, 5.8997, 1.0],
+                           a_max = [1.5, 1.5, 1.5, 1.5, 1.5, 0.5])
+const ptp_path = ptp_path3
 
 
 plotPath(ptp_path, figure=1, ntime=501)
@@ -159,14 +167,14 @@ end
     frame_b = Object3D(rev_b, r=r_rev_b, visualizeFrame=false)
 end
 
-@assembly Gripper(frame_a, right_finger_offset=0.019,
+@assembly Gripper(frame_a, right_finger_offset=0.0,   #  right_finger_offset=0.019
                   d=0.0, k1=50.0, k2=50.0, T2=0.002, gearRatio=1.0) begin
     frame1                   = Object3D(frame_a, R=ModiaMath.rot3(-180u"°"))
     gripper_base_frame       = Object3D(frame1,Solid(SolidFileMesh(gripper_base_frame_obj),0.199,vmat1), visualizeFrame=false)
     gripper_left_finger_a    = Object3D(visualizeFrame=true)
-    gripper_left_finger      = Object3D(gripper_left_finger_a, Solid(SolidFileMesh(gripper_left_finger_obj),0.010,vmat1,contactMaterial="DryWood"), r=[0, 0.0082,0])
+    gripper_left_finger      = Object3D(gripper_left_finger_a, Solid(SolidFileMesh(gripper_left_finger_obj),0.010,vmat1,contactMaterial="DryWood"), r=[0, 0.0082,0])    # 0.0082
     gripper_right_finger_a   = Object3D(gripper_base_frame, r=[0,-right_finger_offset,0])
-    gripper_right_finger     = Object3D(gripper_right_finger_a, Solid(SolidFileMesh(gripper_right_finger_obj),0.010,vmat1,contactMaterial="DryWood"), r=[0,-0.0082,0])
[…]
