### 2019-07-25 e32a5d2
connect(..) returns the constructed joint.
