### 2019-07-28 c61034d
New joint Fixed added together with a test
