### 2019-07-28 76ec3f0
Typo fixed in documentation of Fixed
