### 2019-07-28 e3f7b72
Tests corrected (regarding contact material), so that all are running again
