### 2019-07-28 e95e20e
New ContactBox3 to test box with 8 contact boxes standing on a table + New example of a rolling ball falling from table (to check normal)
