+++ src/contactDetection/ContactDetectionMPR/_module.jl
@@ -21,6 +21,7 @@ import Modia3D.Basics
 import Modia3D.Composition
 import Modia3D.Solids
 import ModiaMath
+import Printf
 
 using StaticArrays
 using DataStructures
