+++ src/Composition/_module.jl
@@ -37,7 +37,7 @@ export updatePosition!, update!, driveJoint!
 export Object3D, AssemblyInternal, initAssemblyInternal!, initAssemblyComponent!, Scene
 export Part
 export SimulationModel
-export BodyWithTwoFrames, ContactBox, ContactBox2
+export BodyWithTwoFrames, ContactBox, ContactBox2, ContactBox3
 export printObject3DAndAllChildren, writeObject3DAndAllChildrenOnJsonFile
 export set_r!, set_q!, set_r_abs!
 export Fixed
