### 2019-07-28 23ed87f
json-files of palettes moved to Modia3D/palettes
