### 2019-07-29 7d8e3c7
No longer used file (contactMaterialPairs.jl) removed
