### 2019-07-31 21a14d9
little changes in final paper
