+++ paper/paper.tex
@@ -17,6 +17,7 @@
 
 \usepackage{booktabs}
 \usepackage{enumitem}
+\usepackage[\sisetup{per-mode=symbol}]{siunitx}
 
 \begin{document}
 
@@ -27,10 +28,11 @@
 \begin{abstract}
 Modia3D is an experimental Julia package to model and simulate 3D mechanical systems.
 Ideas from modern game engines are used to achieve a highly flexible setup
-and features of multi-body programs are used to get a rigid mathematical formulation and
+and features of multi-body algorithms are used to get a rigid mathematical formulation and
 support, for example, of closed kinematic loops. Collision handling is performed on convex
-geometries with elastic response calculation. A Modia3D model is solved with
-a variable-step integrator. This is important to combine Modia3D with the equation-based modeling system Modia in the future.
+geometries with elastic response calculation. 
+A Modia3D model is solved with
+a variable-step solver. This is important to combine Modia3D with the equation-based modeling system Modia in the future.
 \end{abstract}
 
 \input{sections/1_introduction.tex}
