+++ paper/sections/1_introduction.tex
@@ -4,12 +4,12 @@
 \section{Introduction}
 
 Modia3D\footnote{\href{https://github.com/ModiaSim/Modia3D.jl}{https://github.com/ModiaSim/Modia3D.jl}} is an experimental modeling and simulation environment for 3D mechanical systems. It is based on the Julia programming language \cite{bezanson2017julia}.
-Ideas from modern game engines are used to achieve a highly flexible setup of mechanical
+Ideas from modern computer game engines are used to achieve a highly flexible setup of mechanical
 systems including collision handling. Other features are utilized from multi-body programs, 
 such as hierarchical structuring, support of closed kinematic loops and elastic response calculation.
 The underlying mathematical formulation are hybrid DAEs (Differential Algebraic Equations) that are solved with the variable-step solver IDA via the Sundials.jl Julia package \cite{Sundials2005,Rackauckas2017}.
-Emphasis is on variable-step solvers because Modia3D shall be combined in the future with 
-equation-based modeling by using Modia3D assemblies as components in the Julia package
+Emphasis is on variable-step solvers because Modia3D shall be combined with 
+equation-based modeling in the future, using Modia3D assemblies as components in the Julia package
 Modia\footnote{\href{https://github.com/ModiaSim/Modia.jl}{https://github.com/ModiaSim/Modia.jl}} \cite{Modia1,Modia2}
 (for example a joint of a Modia3D system is driven by a Modia model of an 
 electrical motor and gearbox).
@@ -20,4 +20,4 @@ Modia3D provides a generic interface to visualize simulation results with differ
 \emph{DLR Visualization}  library\footnote{\href{http://www.systemcontrolinnovationlab.de/the-dlr-visualization-library/}{http://www.systemcontrolinnovationlab.de/the-dlr-visualization-library/}} \cite{bellmann2009, hellerer2014} are supported. Currently, another team is developing a free 2D/3D web-based authoring tool that will include result visualization.
 
 The user's view of Modia3D was introduced in \cite{Neumayr2018} showing the very flexible definition of 3D systems.
-Some key algorithms are discussed in \cite{Neumayr2017,Neumayr2019}. This article gives an overview of Modia3D from a user's perspective, and in particular how collisions between objects are defined. Furthermore, existing elastic response formulations are combined and enhanced so that a minimum number of material data has to be provided.
+Some key algorithms are discussed in \cite{Neumayr2017,Neumayr2019}. This article gives an overview of Modia3D from a user's perspective, and in particular how collisions between objects are defined. Furthermore, existing elastic response formulations are combined and enhanced so that a minimum number of material data of colliding objects has to be provided.
