+++ paper/sections/5_conclusion.tex
@@ -3,14 +3,14 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 \section{Conclusion}\label{sec:conclusion}
 In this article, a short overview about the experimental 3D modeling environment Modia3D is given.
-In particular, collision handling with a variable-step integrator has been sketched
+In particular, collision handling with a variable-step solver has been sketched
 and a novel formulation for elastic response calculation is proposed.
 The Modia3D collision and contact handling is demonstrated with several examples. 
 
-Modia3D combines ideas from different communities: The architecture with component-oriented modeling
+Modia3D combines ideas from different communities. The architecture with component-oriented modeling
 is inspired by game engines so that 3D models can be setup in a very flexible way, as well as several elements for collision handling.
-Other features are from multi-body programs, like hierarchical structuring, support of closed kinematic loops, and in general to compute results that are close to physics.
+Other features are from multi-body programs, like hierarchical structuring, support of closed kinematic loops, and algorithms to compute results close to real physics.
 
-Modia3D is still a prototype implementation and several important parts are under development, especially the integration with Modia is missing. Furthermore, the code was currently mainly developed for its functionality and is not yet tuned for efficiency. For these reasons,
+Modia3D is still a prototype implementation and several important parts are under development. Especially, the integration with Modia is missing at the moment. Furthermore, the code was currently mainly developed for its functionality and is not yet tuned for efficiency. For these reasons,
 benchmarks and comparisons with other programs with respect to
-simulation efficiency have not been performed yet.
+simulation efficiency have not yet been performed.
