+++ paper/ref.bib
@@ -1,4 +1,4 @@
-﻿@article{bezanson2017julia,
+@article{bezanson2017julia,
   title={{Julia: A fresh approach to numerical computing}},
   author={Bezanson, J. and Edelman, A. and Karpinski, S. and Shah, V. B.},
   journal={SIAM review},
@@ -183,6 +183,8 @@
 	author =	{C. Rackauckas and Q. Nie},
 	title =	{{DifferentialEquations.jl – A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia}},
     journal =	{Journal of Open Research Software},
+    volume={5},
+    number={1},
 	publisher = {Ubiquity Press},
 	doi = { 10.5334/jors.151},
 	year =	{2017},

[…]
