### 2019-08-02 6d316f8
simplification of zero crossing functions (two) one from contact - contact and one from no contact - contact
