+++ examples/collisions/Simulate_NewtonsCradle.jl
@@ -13,7 +13,8 @@ vmatSolids   = Material(color="Red"      , transparency=0.0)   # material of sol
 
 diameter   = 0.02
 lengthWire = 0.15
-rsmall     = diameter/100
+# rsmall     = diameter/100
+rsmall = 0.0
 
 Lx = diameter/2
 Ly = 5*diameter
@@ -46,14 +47,14 @@ end
   pendulum5 = Pendulum()
 
   rev1 = Revolute(frame1, pendulum1.frame1; axis = 1 , phi_start =  -pi/3)
-  rev2 = Revolute(frame2, pendulum2.frame1; axis = 1 )#, phi_start =  -pi/3)
+  rev2 = Revolute(frame2, pendulum2.frame1; axis = 1 )#,phi_start =  -pi/3)
   rev3 = Revolute(frame3, pendulum3.frame1; axis = 1 ) #, phi_start =  -pi/3)
-  rev4 = Revolute(frame4, pendulum4.frame1; axis = 1 )#, phi_start =  pi/3 ) #
-  rev5 = Revolute(frame5, pendulum5.frame1; axis = 1 )# , phi_start =  pi/3 ) #
+  rev4 = Revolute(frame4, pendulum4.frame1; axis = 1 )#,phi_start =  pi/3 ) #
+  rev5 = Revolute(frame5, pendulum5.frame1; axis = 1 )#,phi_start =  pi/3 ) #
 end
 
 gravField     = UniformGravityField(g=9.81, n=[0,0,-1])
-newtonsCradle = NewtonsCradle(sceneOptions=SceneOptions(gravityField=gravField, visualizeFrames=false,
+newtonsCradle = NewtonsCradle(sceneOptions=SceneOptions(elasticContactReductionFactor=1e8, gravityField=gravField, visualizeFrames=false,
                                                         defaultFrameLength=2*diameter, enableContactDetection=true))
 
 # Modia3D.visualizeAssembly!( newtonsCradle )
