+++ src/contactDetection/ContactDetectionMPR/handler.jl
@@ -42,8 +42,7 @@ end
 # This function is called before every integrator step.
 function Composition.selectContactPairsWithEvent!(sim::Union{ModiaMath.SimulationState,NOTHING}, ch::Composition.ContactDetectionMPR_handler, world::Composition.Object3D)
     empty!(ch.contactDict)
-    empty!(ch.noContactDict)
-    empty!(ch.distanceSet)
+    ch.noContactMinVal =  42.0
     selectContactPairs!(sim,ch,world,true)
     ch.contactPairs.nzContact = length(ch.contactDict)
 end
@@ -58,8 +57,7 @@ end
 # It also checks if the contact pairs (contact == true), from a previous call of selectContactPairsWithEvent!(...)
 # are still true.
 function Composition.selectContactPairsNoEvent!(sim::Union{ModiaMath.SimulationState,NOTHING}, ch::Composition.ContactDetectionMPR_handler, world::Composition.Object3D)
-    empty!(ch.noContactDict)
-    empty!(ch.distanceSet)
+    ch.noContactMinVal =  42.0
     selectContactPairs!(sim,ch,world,false)
 end
 
@@ -72,24 +70,15 @@ function selectContactPairs!(sim::Union{ModiaMath.SimulationState,NOTHING}, ch::
   # Store z in simulation engine
   if typeof(sim) != Nothing
     simh = sim.eventHandler
-    i = 0
-
-    for (key, pair) in ch.contactDict
-      i = i + 1
-      simh.z[i] = pair.distanceWithHysteresis
+    # z[1] ... zero crossing function from contact to no contact
+    if isempty(ch.contactDict)
+      simh.z[1] = -42.0
+    else
+      (pair, key)  = findmax(ch.contactDict)
+      simh.z[1] = pair.distanceWithHysteresis
     end
-
-
-    for (key, pair) in ch.noContactDict
-      i = i + 1
-      simh.z[i] = pair.distanceWithHysteresis
-    end
-
-    # Provide dummy value for not used z-entry
-    while i < ch.contactPairs.nz
-        i = i + 1
-        simh.z[i] = 42.0
-    end
+    # z[2] ... zero crossing function from no contact to contact
+    simh.z[2] = ch.noContactMinVal
   end
 
   ch.distanceComputed = true
@@ -151,40 +140,25 @@ function computeDistances(ch::Composition.ContactDetectionMPR_handler, world::Co
 end
 
 
-const zEps  = 1.e-8
+const zEps  = 1.e-10
 const zEps2 = 2*zEps
 
 
-function pushCollisionPair!(ch, contact, distanceWithHysteresis, pairID, contactPoint1,contactPoint2,contactNormal,actObj,nextObj,hasevent)::Nothing
-    distanceKey = Composition.DistanceKey(contact, distanceWithHysteresis, pairID)
-    if  hasevent && (length(ch.distanceSet) < ch.contactPairs.nz) ||
-       !hasevent && (length(ch.distanceSet) + length(ch.contactDict) < ch.contactPairs.nz)
-        push!(ch.distanceSet, distanceKey)
-        if contact
-            push!(ch.contactDict  , pairID => Modia3D.ContactPair(  contactPoint1,contactPoint2,contactNormal,actObj,nextObj,distanceWithHysteresis))
-        else
-            push!(ch.noContactDict, pairID => Modia3D.NoContactPair(contactPoint1,contactPoint2,contactNormal,actObj,nextObj,distanceWithHysteresis))
-        end
+function pushCollisionPair!(ch, contact, distanceWithHysteresis, pairID,
+                            contactPoint1,contactPoint2,contactNormal,actObj,nextObj,hasevent)::Nothing
+    if contact
[…]
