### 2019-08-02 989bc46
Documentation docs Internal adapted to new implementation of crossing function + docu added to example Simulate_NewtonsCradle
