+++ src/Solids/contactMaterial.jl
@@ -1,4 +1,4 @@
-# License for this file: MIT (expat)
+﻿# License for this file: MIT (expat)
 # Copyright 2017-2018, DLR Institute of System Dynamics and Control
 #
 # This file is part of module
@@ -37,184 +37,6 @@ Available keys are for example `"Steel", "Aluminium", "DryWood", "BilliardBall"`
 - `cor::Float64`: Coefficient of restitution between *two objects* (``0 \\le cor \\le 1``).
 - `mu_k::Float64`: Kinetic/sliding friction force coefficient between *two objects* (``\\ge 0``).
 - `mu_r::Float64`: Rotational rolling resistance torque coefficient between *two objects* (``\\ge 0``).
-
-
-# Mathematical description of the response calculation
-
-When two 3D objects penetrate each other with a penetration depth ``\\delta > 0``
-then a *contact force* and a *contact torque* is computed from the elastic contact materials
-of the two objects in the following way
-(the contact force law in normal direction is based on [^1], [^3],
-the remaining force law on [^2] with some extensions and corrections):
-
-
-```math
-\\begin{align}
-f_n        &= \\max\\left(0, c_{res} \\cdot c_{geo} \\cdot \\delta^{n_{geo}} \\cdot (1 + d \\cdot \\dot{\\delta}) \\right) \\\\
-\\vec{f}_n &= f_n \\cdot \\vec{e}_n \\\\
-\\vec{f}_t &= -\\mu_{k} \\cdot f_n \\cdot \\vec{e}_{t,reg} \\\\
-\\vec{\\tau} &= -\\mu_{r} \\cdot \\mu_{r,geo}  \\cdot f_n \\cdot \\vec{e}_{\\omega,reg}
-\\end{align}
-```
-
-where
-
-- ``\\vec{e}_n``: Unit vector perpendicular to the surface of object 1 at the contact point, pointing outwards.
-- ``\\vec{e}_{t,reg}``: Regularized unit vector in direction of the tangential relative velocity (see below).
-- ``\\vec{e}_{\\omega,reg}``: Regularized unit vector in direction of the relative angular velocity (see below).
-- ``f_n``: Value of normal contact force in direction of ``\\vec{e}_n`` acting on object 2 (``f_n \\ge 0``).
-- ``\\vec{f}_n``: Vector of normal contact force acting on object 2.
-- ``\\vec{f}_t``: Vector of sliding friction force acting on object 2 in opposite direction of the movement
-                  in the tangential plane of the contact.
-- ``\\vec{\\tau}``: Vector of rolling resistance torque acting on object 2, in opposite direction to the
-                    relative angular velocity between the two contacting objects.
-- ``\\delta``: Signed distance between object 1 and 2 in normal direction ``\\vec{e}_n``.
-               ``\\delta > 0`` if objects are penetrating each other.
-- ``\\dot{\\delta}``: Signed relative velocity between object 1 and 2 in normal direction ``\\vec{e}_n``.
-- ``c_{res}``: Resultant elastic material constant in normal direction. This constant is computed
-         from the constants ``c_1, c_2`` of the two contacting objects 1 and 2 as
-         ``c_{res} = 1/\\left( 1/c_1 + 1/c_2 \\right)``. ``c_i`` is computed from the material properties as
-         ``c_i = E_i/(1 - \\nu_i^2)`` where ``E_i`` is Young's modules and
-         ``\\nu_i`` is Poisson's ratio of object i.
-- ``c_{geo}``: Factor in ``f_n`` that is determined from the geometries of the
-               two objects (see below).
-- ``n_{geo}``: Exponent in ``f_n`` that is determined from the geometries of the
-               two objects (see below).
[…]
