+++ docs/src/internal/ContactDetection.md
@@ -7,8 +7,6 @@ In `scene.options.contactDetection`
 (file `Modia3D/src/contactDetection/ContactDectionMPR_handler.jl`)
 information about the contact situation is stored in an instance
 of [`Modia3D.Composition.ContactDetectionMPR_handler`](@ref).
-Hereby information about at most `nz_max` objects pairs are stored that
-are in contact or are close to each other.
 
 The *distance* between two objects is defined by two variables:
 
@@ -61,156 +59,39 @@ identical to zero when starting after initialization or restarting after an
 event.
 
 
-## Keys
+## Crossing functions
 
-An *object pair* is identified by a *unique* `pairID::PairID`, where `const PairID = Int64`.
-A pairID is constructed by packing the object indices of the two objects into an `Int64` number.
+Modia3D uses the dictionary `contactDict` to keep track of the
+contact situation. Every pair of objects is identified by a unique
+Integer value called *PairID* that is used as key in `conctactDict`.
+A dictionary value is an instance of [`Modia3D.Composition.ContactPair`](@ref)
 
-The information about the *smallest distances* are stored in the sorted
-dictionary `distanceSet` (see below).
-The entries in this dictionary are ordered according to their `key::DistanceKey`.
-This key consists of three elements that are evaluated in the following order:
+At an event instant, dictionary `contactDict` is emptied and all object pairs are stored
+newly in `contactDict` that have `distanceWithHysteresis < 0`.
+Between events, contact forces/torques are only
+applied on object pairs that are in this dictionary.
 
-1. `contact`
-2. `distanceWithHysteresis`
-3. `pairID`
+The following functions are used as zero crossing functions:
 
-If `contact=true`, order does not matter and is therefore only based on the unique `pairID`. If `contact=false`, order matters and is basically on `distanceWithHysteresis`. This leads to the following definition of the `DistanceKey` ordering:
+* ``z_1(t)``: The maximum `distanceWithHysteresis` of all object pairs that are in `contactDict`.
+* ``z_2(t)``: The minimum `distanceWithHysteresis` of all object pairs that are **not** in `contactDict`.
 
-```
-Base.:isless(key1::DistanceKey, key2::DistanceKey) =
-     key1.contact &&  key2.contact ? key1.pairID < key2.pairID :
-   ( key1.contact && !key2.contact ? true                        :
-   (!key1.contact &&  key2.contact ? false                       :
-   (key1.distanceWithHysteresis != key2.distanceWithHysteresis   ?
-    key1.distanceWithHysteresis < key2.distanceWithHysteresis    :
-    key1.pairID                < key2.pairID)))
-```
+As a result ``z_1`` monitors if an object pair that has been in contact, looses contact, so
+`distanceWithHysteresis` becomes greater than zero,
+and ``z_2`` monitors if an object pair that has not been in contact, starts to penetrate each other, so
+`distanceWithHysteresis` becomes less than zero,
 
+Whenever the integrator requires the values of the zero crossing functions, the values of
+``z_1, z_2`` are newly computed. At all other time instants (e.g. communication points),
+only the distances of the object pairs are computed that are in dictionary `contactDict`,
+because these distances are used to compute the contact forces and torques. Distances
+of object pairs that are not in contact to each other are not computed at these time instant.
 
-## Dictionaries
-
-The following dictionaries are utilized for contact detection:
-
-| Dictionary/Set   | key type    | value type    | contact    | step     | else   |
-|:---------------- |:------------|:--------------|:----------:|:--------:|:------:|
-|`lastContactDict` | PairID      | ContactPair   | true       |   -      |   -    |
-|`contactDict`     | PairID      | ContactPair   | true       | update   | update |
[…]
