### 2019-08-26 #30 @matbesancon
@ChrisRackauckas would you like to review this paper?
