### 2019-08-26 #30 @matbesancon
@MartinOtter sorry about the delay on this paper. @Datseris I will find a second reviewer but you can start reviewing the paper and code already (you can see example checklists on the other issues)
