### 2019-08-27 #30 @Datseris
@matbesancon there is no review issue open on this. Having the checklist is pretty helpful, can you please start the review? In another paper that I reviewed this is how it progressed and a second reviewer was added after I have finished my review.
