### 2019-08-27 #43 @whedon
[REVIEW]: Modia3D: Modeling and Simulation of 3D-Systems in Julia
**Submitting author:** @MartinOtter (<a href="http://orcid.org/0000-0001-6348-9569">Martin Otter</a>)
**Repository:** <a href="https://github.com/ModiaSim/Modia3D.jl" target ="_blank">https://github.com/ModiaSim/Modia3D.jl</a>
