### 2019-08-27 #43 @matbesancon
@dawbarton would you be available to review this paper?
