### 2019-08-27 #43 @matbesancon
@MartinOtter I think you need `ref.bib` in the paper.yml file
