### 2019-09-03 #43 @matbesancon
ping @MartinOtter to get the paper compiling
