### 2019-09-03 16facee
Latex package definition corrected + paper.pdf updated
