### 2019-09-03 #43 @MartinOtter
The issue has now been corrected and latexmk should generate a pdf file (seems to be that latexmk did not like the option of a \usepackage{siunitx}).

Note, a pdf-version is also available as [https://github.com/ModiaSim/Modia3D.jl/blob/JuliaCon2019-paper/paper.pdf](https://github.com/ModiaSim/Modia3D.jl/blob/JuliaCon2019-paper/paper.pdf).
