### 2019-09-03 #43 @matbesancon
@MartinOtter siunitx.sty seems to be missing, you can use the command i did above to try to compile your paper once it is done
