### 2019-09-03 b4eff15
siunitx package explicitly added
