### 2019-09-03 #43 @MartinOtter
Added siunitx.sty package explicitly
