### 2019-09-05 #43 @matbesancon
@MartinOtter  can you compile the paper when you change something?
