### 2019-09-05 #43 @matbesancon
@Datseris, @dawbarton I think you can continue the review here
