### 2019-09-05 #43 @MartinOtter
>
>
> @MartinOtter can you compile the paper when you change something?

Yes:

I clone the branch to a local directory and use the command `latexmk -bibtex -pdf paper.tex` to generate a `paper.pdf` file. I run this on Windows 7, after installing Ruby and latexmk.
