### 2019-09-07 #43 @Datseris
Hi @matbesancon  and @MartinOtter , I would like to state my first round of review.

The installation of the software proceed as planned, and I can confirm that tests pass on my machine (Windows 10, Julia 1.2). I must disclose that **I will not be testing nor reviewing any of the plotting claims of the software**. This requires you to (1) install a 3rd party software, completely unrelated with the Julia language and (2) share with them your contact details (!?!?).

The paper describing the software is very well written. It has multiple examples and goes in depth in the algorithms used and the mathematics that accompany them. The authors also take care how the Julia programming language is taken advantage of in their work. Truthfully, I was very pleased to see such a level of detail and care taken for this paper. Unfortunately, the online documentation is not yet at this level.

---

Here are my comments:

1. In the paper and documentation there is a lack of comparison with other software that achieve similar characteristics. The software should be put into context of other existing software, and also attempt to educate the potential user on why they should choose this software, instead of other. I find it hard to believe that Modia3D.jl is the only software that does what it does.
2. In the start of paragraph 2 the authors "compare" Modia3D.jl with game engines, but this is not a true comparison. They only mention that they use game developing paradigms without going into much depth. Me, as a potential user, should I consider comparing your software with Unreal Engine or with DifferentialEquations.jl ?
2. I think the authors assume knowledge of what Modia or Modelica is. The reader of the paper may not necessarily have this knowledge (I for example do not). I think the authors should try to describe Modia in a couple of sentences before using it in text.
3. Both the paper as well as the documentation lack a statement of need. This also ties with my first comment; the authors do not really put the work into context.
4. There is no discussion about performance but I think the authors should consider providing one. Does this software match the performance of its competitors? Does it exceed it? Furthermore, novel algorithms seem to have been developed and be in use for this software (the three papers associated with the software, e.g. *"Collision Handling with Variable-Step Integrators (EOOLT 2017, December)"*. Do these approaches also offer performance benefits?
5. The documentation is severely lacking examples as well as instructions. The package features are listed in detail, but instructions of how to leverage them are nowhere to be seen. Specifically, the "getting started page" is simply installation instructions, while all the "examples" of the software are presented like so:
    ```
    # Examples to demonstrate kinematic movement of assemblies
    include("$(Modia3D.path)/examples/kinematics/Move_DoublePendulum.jl")
    include("$(Modia3D.path)/examples/kinematics/Move_FourBar.jl")
    include("$(Modia3D.path)/examples/kinematics/Move_FourBar2.jl")
    ```
   This is not enough. At least one example must be expanded in detail and every single step of the example must be explained, also in detail. In fact, at the moment I cannot see how a user would actually use Modia3D.jl, given the absence of guidance. Please be aware that listing an example is not enough: you rely on me being able to *deduce* what a command does by seeing its output. This is not the proper way to learn, as I might deduce many things wrongly.
6. Although optional, I think the authors should strongly consider bundling this plotting library they use with their software. The current situation is a no-go for me.

---

@matbesancon I think you must clarify the following, because a lot of my comments concern documentation and being able to use the software. In many places the authors state that this is "an experimental software" and "not yet ready for production". And in fact the software version is 0.3.0. These statements already justify the lack of documentation.

Please clarify what am I supposed to review here: am I reviewing a paper that explains novel algorithms or am I reviewing a software that I could be potentially using at my work?
