### 2019-09-19 c32c56b
start with cleaning up on test examples for collision
