+++ test/test_Examples_Collision.jl
@@ -4,6 +4,17 @@ import Modia3D
 
 # Test Modia3D - Examples
 
+# test/collision
+
+collisionPath = joinpath(Modia3D.path, "test", "collision")
+# include(joinpath(collisionPath, "Collision_3Elements.jl"))
+# include(joinpath(collisionPath, "Collision_Bars.jl")) # is not working needs to be fixed
+include(joinpath(collisionPath, "Test_Collision.jl"))
+include(joinpath(collisionPath, "Test_Collision_moreRevolutes.jl"))
+include(joinpath(collisionPath, "Test_Collision_StarSetting.jl"))
+include(joinpath(collisionPath, "Test_MiniBsp.jl"))
+include(joinpath(collisionPath, "Test_Solids.jl"))
+
 # test/dynamics
 dynamicsPath = joinpath(Modia3D.path, "test", "dynamics")
 include(joinpath(dynamicsPath, "Simulate_ControllerDamper.jl"))
