### 2019-09-19 #43 @matbesancon
> In many places the authors state that this is "an experimental software" and "not yet ready for production". And in fact the software version is 0.3.0. These statements already justify the lack of documentation.

Having accessible documentation and software that can be used by people other than the main developer(s) is in the criteria for the proceedings.

> am I reviewing a paper that explains novel algorithms or am I reviewing a software that I could be potentially using at my work?

We are reviewing both, in the end, the proceedings will store the paper and the DOI corresponding to the version of the software
