### 2019-09-19 #43 @matbesancon
@dawbarton you didn't seem to have a checklist so I added it on the first whedon comment
