### 2019-09-19 #43 @matbesancon
It would be nice to wrap things up before here then
