### 2019-09-23 487dbab
clean up of collision test 1
