+++ test/collision/dynamic_collision/collision_BallWithBox_Prismatic.jl
@@ -1,4 +1,4 @@
-module collision_Billard
+module collision_BallWithBox_Prismatic
 
 using Modia3D
 using Modia3D.StaticArrays
@@ -7,9 +7,8 @@ import Modia3D.ModiaMath
 
 vmatGraphics = Modia3D.Material(color="LightBlue" , transparency=0.5)    # material of Graphics
 vmatSolids = Modia3D.Material(color="Red" , transparency=0.5)         # material of solids
-#c = 1e9, d = 100.0
-cmat = Modia3D.ContactMaterialElastic(c=1e9, d=0.0)
 
+cmat = "Steel"
 
 LxGround = 16.0
 LyGround = 8.0
@@ -21,26 +20,19 @@ diameter = 1.0
 
 @assembly Billard() begin
   world = Modia3D.Object3D(visualizeFrame=true)
-  lengthBox = Modia3D.Solid(Modia3D.SolidBox(LyBox, LyGround, LzBox) , "Aluminium", vmatSolids; contactMaterial = cmat)
+  lengthBox = Modia3D.Solid(Modia3D.SolidBox(LyBox, LyGround, LzBox) , "Steel", vmatSolids; contactMaterial = cmat)
   box4 = Modia3D.Object3D(world, lengthBox, r=[-LxGround/2, 0.0, LzGround], fixed=true)
 
-  sphere = Modia3D.Object3D( Modia3D.Solid(Modia3D.SolidSphere(diameter) , "Aluminium" , vmatSolids ; contactMaterial = cmat))
+  sphere = Modia3D.Object3D( Modia3D.Solid(Modia3D.SolidSphere(diameter) , "Steel" , vmatSolids ; contactMaterial = cmat))
   prisX = Modia3D.Prismatic(world, sphere; axis=1, canCollide = true)
-
-
-
 end
-
 
 gravField = Modia3D.UniformGravityField(g=9.81, n=[-1,0,0])
 bill = Billard(sceneOptions=Modia3D.SceneOptions(gravityField=gravField,visualizeFrames=true, defaultFrameLength=0.7,nz_max = 100, enableContactDetection=true, visualizeContactPoints=true, visualizeSupportPoints=true))
-
-# Modia3D.visualizeAssembly!( bill )
-
 
 model = Modia3D.SimulationModel( bill )
 result = ModiaMath.simulate!(model; stopTime=1.5, tolerance=1e-8,interval=0.001, log=false)
 
 
-println("... success of collision_Billard.jl!")
+println("... success of collision_BallWithBox_Prismatic.jl!")
 end

[…]
