+++ README.md
@@ -11,7 +11,7 @@
 Modia3D is a Julia package to model fixed and moving objects in 3D (*e.g.* visual shapes, rigid bodies).
 These objects are driven kinematically by pre-defined time functions or are moving dynamically by
 solving Differential Algebraic Equations (DAEs)
-with a variable-step DAE solver. 
+with a variable-step DAE solver.
 
 Collision handling with elastic response calculation is
 performed for objects that are defined with a contact material and (a) have a convex geometry,
@@ -22,8 +22,8 @@ A more detailed overview of the available features is given in the
 Papers about Modia3D:
 
 - *[Collision Handling with Variable-Step Integrators](docs/resources/documentation/CollisionHandling_Neumayr_Otter_2017.pdf)* ([EOOLT 2017, December](http://www.eoolt.org/2017/))
-- *[Component-Based 3D Modeling of Dynamic Systems](http://www.ep.liu.se/ecp/154/019/ecp18154019.pdf)*([American Modelica Conference 2018, October](https://www.modelica.org/events/modelica2018Americas/index_html))
-- *[Algorithms for Component-Based 3D Modeling](http://www.ep.liu.se/ecp/157/039/ecp19157039.pdf)*([13th International Modelica Conference 2019, March](https://modelica.org/events/modelica2019))
+- *[Component-Based 3D Modeling of Dynamic Systems](http://www.ep.liu.se/ecp/154/019/ecp18154019.pdf)* ([American Modelica Conference 2018, October](https://www.modelica.org/events/modelica2018Americas/index_html))
+- *[Algorithms for Component-Based 3D Modeling](http://www.ep.liu.se/ecp/157/039/ecp19157039.pdf)* ([13th International Modelica Conference 2019, March](https://modelica.org/events/modelica2019))
 
 
 Before releasing version 1.0, Modia3D shall be
@@ -31,6 +31,13 @@ easily combinable with [Modia](https://github.com/ModiaSim/Modia.jl), for exampl
 electrical motor with Modia, and add 3D behavior/visualization with Modia3D.
 By this approach the best of both worlds can be combined:
 Special 3D algorithms (Modia3D) + power/flexibility of equation based modeling (Modia).
+
+
+## Modia Project
+
+The [Modelica standard library](https://github.com/modelica/ModelicaStandardLibrary) supports the modeling of 3-dimensional multi-body systems with its sub library Modelica.Mechanics.MultiBody. There have been several attempts to improve this library with regards to visualization, collision handling or support of larger models, for example. Over the years it was recognized that the technology of current [Modelica](https://www.modelica.org) has some natural limitations.
+
+Therefore, the open source [Modia](https://github.com/ModiaSim/Modia.jl) project was launched as a domain-specific extension of the  [Julia](https://julialang.org) programming language. The equation-based modeling language is called Modia as well as the experimental [modeling environment](https://github.com/ModiaSim) consisting of several Julia packages. The intention is to utilize the results of this prototyping in the design of the next Modelica language generation.
 
 
 ## Installation
