### 2019-09-24 8eddd9a
update of collision tests
