+++ test/collision/contactForceLaw/contactForceLaw_newtons_cradle.jl
@@ -1,4 +1,4 @@
-module collision_newtons_cradle
+module contactForceLaw_newtons_cradle
 
 using Modia3D
 using Modia3D.StaticArrays
@@ -8,19 +8,14 @@ import Modia3D.ModiaMath
 vmatGraphics = Modia3D.Material(color="LightBlue" , transparency=0.1)    # material of Graphics
 vmatSolids = Modia3D.Material(color="Red" , transparency=0.0)         # material of solids
 
-
-cmat = Modia3D.ElasticContactMaterial2("BilliardBall")
+cmat = "Steel"
 
 Lx = 0.3
 Ly = 4.1
 Lz = 0.1
 
-
-
 diameter = 1.0
 lengthWire = 4.0
-
-
 
 @assembly NewtonComparison begin
   world = Modia3D.Object3D(visualizeFrame=false)
@@ -28,19 +23,14 @@ lengthWire = 4.0
 #  cradle2 = NewtonsCradleSolidWire(world)
 end
 
-
-
 @assembly NewtonsCradle(world) begin
-#   world = Modia3D.Object3D(visualizeFrame=false)
   box    = Modia3D.Object3D(world, Modia3D.Box(Lx, Ly, Lz; material=vmatGraphics)   ; r=[0.0, 0.0, 0.0] )
 
   frame1 = Modia3D.Object3D(box; r=[0.0, -Ly/2, 0.0])
-
   frame2 = Modia3D.Object3D(box; r=[0.0, -Ly/4, 0.0])
   frame3 = Modia3D.Object3D(box; r=[0.0, 0.0, 0.0])
   frame4 = Modia3D.Object3D(box; r=[0.0, Ly/4, 0.0])
   frame5 = Modia3D.Object3D(box; r=[0.0, Ly/2, 0.0])
-
 
   pendulum1 = Pendulum()
   pendulum2 = Pendulum()
@@ -67,7 +57,6 @@ end
   sphere = Modia3D.Object3D(frame2, Modia3D.Solid(Modia3D.SolidSphere(diameter) , "Steel", vmatSolids; contactMaterial = cmat) )
 end
 
-
 @assembly NewtonsCradleSolidWire(world) begin
   # world = Modia3D.Object3D(visualizeFrame=false)
   box    = Modia3D.Object3D(world, Modia3D.Box(Lx, Ly, Lz; material=vmatGraphics); r=[0.0, -8.0, 0.0] )
@@ -93,7 +82,6 @@ end
   rev5   = Modia3D.Revolute(frame5, pendulum5.frame1; axis = 1  , phi_start =  pi/3 ) #
 end
 
-
 @assembly Pendulum1() begin
   wire = Modia3D.Object3D(Modia3D.Solid(Modia3D.SolidCylinder(diameter/5, lengthWire), "Steel", vmatSolids) )
 
@@ -109,12 +97,8 @@ end
 gravField = Modia3D.UniformGravityField(g=9.81, n=[0,0,-1])
 newton = NewtonComparison(sceneOptions=Modia3D.SceneOptions(gravityField=gravField,visualizeFrames=false, defaultFrameLength=0.7,nz_max = 100, enableContactDetection=true, visualizeContactPoints=false, visualizeSupportPoints=false))
 
-# Modia3D.visualizeAssembly!( newton )
-
-
 model = Modia3D.SimulationModel( newton )
 result = ModiaMath.simulate!(model; stopTime=10.0, tolerance=1e-8,interval=0.001, log=false)
 
-
-println("... success of collision_newtons_cradle.jl!")
+println("... success of contactForceLaw_newtons_cradle.jl!")
 end
