+++ test/collision/Billard_Simulations/rollingBall/rollingBall1.jl
@@ -1,4 +1,4 @@
-module contactForceLaw_rollingBall_1ball
+module rollingBall1
 
 using Modia3D
 using Modia3D.StaticArrays
@@ -9,46 +9,34 @@ vmatGraphics = Modia3D.Material(color="LightBlue" , transparency=0.5)    # mater
 vmatSolids = Modia3D.Material(color="Red" , transparency=0.0)         # material of solids
 vmatTable = Modia3D.Material(color="Green", transparency=0.1)         # material of table
 
-cmatTable = Modia3D.ElasticContactMaterial2("BilliardTable")
-cmatBall = Modia3D.ElasticContactMaterial2("BilliardBall")
+cmatTable = "BilliardTable"
+cmatBall = "BilliardBall"
 
 LxGround = 3.0
-
-
-
 LyBox = 0.5
 LzBox = 0.02
 diameter = 0.06
 @assembly Table(world) begin
-  withBox = Modia3D.Solid(Modia3D.SolidBox(LxGround, LyBox, LzBox) , "DryWood", vmatTable; contactMaterial = cmatTable)
-  box1 = Modia3D.Object3D(world, withBox, r=[1.5, 0.0, -LzBox/2], fixed=true, visualizeFrame=false)
+  withBox = Modia3D.Solid(Modia3D.SolidBox(LxGround, LyBox, LzBox) , "BilliardTable", vmatTable; contactMaterial = cmatTable)
+  box1 = Modia3D.Object3D(world, withBox, r=[1.5, 0.0, -LzBox/2], fixed=true)
 end
-
-sphere = Modia3D.Solid(Modia3D.SolidSphere(diameter), "BilliardBall", vmatSolids ;   contactMaterial = cmatBall)
 
 @assembly OneRollingBall() begin
   world = Modia3D.Object3D(visualizeFrame=false)
   table = Table(world)
-  ball1 = Modia3D.Object3D(world, sphere, fixed = false, r=[0.2, 0.0, diameter/2], v_start=[3.0, 0.0, 0.0], visualizeFrame=true )
+  ball1 = Modia3D.Object3D(world, Modia3D.Solid(Modia3D.SolidSphere(diameter), "BilliardBall", vmatSolids ;
+                           contactMaterial = cmatBall), fixed = false, r=[0.2, 0.0, diameter/2], v_start=[3.0, 0.0, 0.0], visualizeFrame=true )
 end
-
 
 gravField = Modia3D.UniformGravityField(g=9.81, n=[0,0,-1])
 bill = OneRollingBall(sceneOptions=Modia3D.SceneOptions(gravityField=gravField,visualizeFrames=false,
-                      defaultFrameLength=0.1,nz_max = 100, enableContactDetection=true, visualizeContactPoints=false, visualizeSupportPoints=false))
-
-
-
-#Modia3D.visualizeAssembly!( bill )
-
+                       defaultFrameLength=0.1,nz_max = 100, enableContactDetection=true, visualizeContactPoints=false, visualizeSupportPoints=false))
 model = Modia3D.SimulationModel( bill )
-#ModiaMath.print_ModelVariables(model)
 
-#=
 using PyPlot
 using PyCall
 
-pyplot_rc = PyCall.PyDict(PyPlot.matplotlib."rcParams")
+pyplot_rc = PyCall.PyDict(PyPlot.matplotlib["rcParams"])
 pyplot_rc["font.family"]      = "sans-serif"
 pyplot_rc["font.sans-serif"]  = ["Calibri", "Arial", "Verdana", "Lucida Grande"]
 pyplot_rc["font.size"]        = 12.0
[…]
