+++ test/runtests.jl
@@ -11,6 +11,7 @@ using Modia3D.Test
    include("test_Composition.jl")
    include("test_DynamicExamples.jl")
    include("test_Examples.jl")
+   include("test_Examples_Collision.jl")
    include("runexamples.jl")
 
    println("\n... success of all tests!")
