+++ README.md
@@ -19,31 +19,35 @@ Therefore, the open source [Modia](https://github.com/ModiaSim/Modia.jl) project
 
 One part of the Modia project is [Modia3D](https://github.com/ModiaSim/Modia3D.jl). Initially, it is an experimental modeling and simulation environment for 3D mechanical systems, but it shall be expanded into other domains in the future.
 
-Modia3D is a Julia package to model fixed and moving objects in 3D (*e.g.* visual shapes, rigid bodies).
-These objects are driven kinematically by pre-defined time functions or are moving dynamically by
-solving Differential Algebraic Equations (DAEs)
-with a variable-step DAE solver.
+Ideas from modern computer game engines are used to achieve a highly flexible setup of mechanical systems including collision handling. Other features are utilized from multi-body programs, such as hierarchical structuring, support for closed kinematic loops, and elastic response calculation.
+The underlying mathematical formulation are hybrid Differential Algebraic Equations (DAEs) that are solved with the variable-step solver IDA via the Sundials.jl Julia package.
+Emphasis is on variable-step solvers to combine Modia3D with equation-based modeling in the future, using Modia3D assemblies as components within one high level programming environment (for example a joint of a Modia3D system is driven by a Modia model of an electrical motor and gearbox).
 
-Collision handling with elastic response calculation is
-performed for objects that are defined with a contact material and (a) have a convex geometry,
-(b) can be approximated by a set of convex geometries, or (c) have a concave geometry
-that is (automatically) approximated by its convex hull.
+Features of the Modia3D Julia package are modeling fixed and moving objects in 3D (*e.g.* visual shapes, rigid bodies). These objects are driven kinematically by pre-defined time functions or are moving dynamically by solving Differential Algebraic Equations (DAEs) with a variable-step DAE solver.
+
+Further, collision handling with elastic response calculation is performed for objects that are defined with a contact material and have a convex geometry, or can be approximated by a set of convex geometries, or have a concave geometry which is (automatically) approximated by its convex hull.
+Collision handling with elastic response calculation and error controlled integration is challenging.
+Contact handling in Modia3D uses variable-step solvers where the penetration depths and Euclidean distances computed with the improved Minkowski Portal Refinement (MPR) algorithm [1] are utilized to construct appropriate zero-crossing functions.
+Whenever a collision with elastic response calculation occurs, the model response is drastically changed. An advanced contact law is introduced in [5] and many detailed examples using it are discussed in [4]. The user's view of Modia3D is introduced in [2] showing the very flexible definition of 3D systems. Some key algorithms are discussed in [3].
 A more detailed overview of the available features is given in the
 [Modia3D documentation](https://ModiaSim.github.io/Modia3D.jl/stable).
-Papers about Modia3D:
+Papers and videos about Modia3D:
 
-- *[Collision Handling with Variable-Step Integrators](docs/resources/documentation/CollisionHandling_Neumayr_Otter_2017.pdf)* ([EOOLT 2017, December](http://www.eoolt.org/2017/))
-- *[Component-Based 3D Modeling of Dynamic Systems](http://www.ep.liu.se/ecp/154/019/ecp18154019.pdf)* ([American Modelica Conference 2018, October](https://www.modelica.org/events/modelica2018Americas/index_html))
[…]
