+++ README.md
@@ -9,21 +9,19 @@
 
 
 ## Modia Project
-
 The [Modelica standard library](https://github.com/modelica/ModelicaStandardLibrary) supports the modeling of 3-dimensional multi-body systems with its sub library Modelica.Mechanics.MultiBody. There have been several attempts to improve this library with regards to visualization, collision handling or support of larger models. Over the years it was recognized that the technology of current [Modelica](https://www.modelica.org) has some natural limitations.
 
 Therefore, the open source [Modia](https://github.com/ModiaSim/Modia.jl) project was launched as a domain-specific extension of the  [Julia](https://julialang.org) programming language. The equation-based modeling language is called Modia as well as the experimental [modeling environment](https://github.com/ModiaSim) consisting of several Julia packages. The intention is to utilize the results of this research and prototyping in the design of the next Modelica language generation.
 
 
 ## Modia3D
-
 One part of the Modia project is [Modia3D](https://github.com/ModiaSim/Modia3D.jl). Initially, it is an experimental modeling and simulation environment for 3D mechanical systems, but it shall be expanded into other domains in the future.
 
 Ideas from modern computer game engines are used to achieve a highly flexible setup of mechanical systems including collision handling. Other features are utilized from multi-body programs, such as hierarchical structuring, support for closed kinematic loops, and elastic response calculation.
-The underlying mathematical formulation are hybrid Differential Algebraic Equations (DAEs) that are solved with the variable-step solver IDA via the Sundials.jl Julia package.
+The underlying mathematical formulation are hybrid Differential Algebraic Equations (DAEs) that are solved with the variable-step solver IDA via the [Sundials.jl](https://github.com/JuliaDiffEq/Sundials.jl) Julia package.
 
 
-Features of the Modia3D Julia package are modeling fixed and moving objects in 3D (*e.g.* visual shapes, rigid bodies). These objects are driven kinematically by pre-defined time functions or are moving dynamically by solving Differential Algebraic Equations (DAEs) with a variable-step DAE solver.
+Features of the Modia3D Julia package are modeling fixed and moving objects in 3D (e.g. visual shapes, rigid bodies). These objects are driven kinematically by pre-defined time functions or are moving dynamically by solving Differential Algebraic Equations (DAEs) with a variable-step DAE solver.
 
 Further, collision handling with elastic response calculation is performed for objects that are defined with a contact material and have a convex geometry, or can be approximated by a set of convex geometries, or have a concave geometry which is (automatically) approximated by its convex hull.
 Collision handling with elastic response calculation and error controlled integration is challenging.
@@ -40,17 +38,14 @@ Papers and videos about Modia3D:
 - [5] *Collision Handling with Elastic Response Calculation and Zero-Crossing Functions* (will be published on ([EOOLT 2019, November](http://www.eoolt.org/2019/))
 
 ### 3D Renderer and Work-In-Progress Web App
-
 Modia3D provides a generic interface to visualize simulation results with different 3D renderers. The [free community edition](https://visualization.ltx.de/) as well as the professional edition of the
[…]
