+++ README.md
@@ -54,41 +54,25 @@ julia> ]add Modia3D
         add PyPlot     # in order that plots are shown
 ```
 
-Modia3D uses [PyPlot](https://github.com/JuliaPy/PyPlot.jl) for plotting.
-If `PyPlot` is not available in your current Julia environment
-an information message is printed and all `plot(..)` calls are ignored.
+Modia3D uses [PyPlot](https://github.com/JuliaPy/PyPlot.jl) for plotting. If `PyPlot` is not available in your current Julia environment an information message is printed and all `plot(..)` calls are ignored.
 
-In order that plot windows are displayed, you need to add `PyPlot` to your current environment
-via `]add PyPlot`. Often this automatic installation fails and it is recommended to follow
-the instructions
-[Installing PyPlot in a robust way](https://github.com/ModiaSim/ModiaMath.jl/wiki/Installing-PyPlot-in-a-robust-way).
+In order that plot windows are displayed, you need to add `PyPlot` to your current environment via `]add PyPlot`. Often this automatic installation fails and it is recommended to follow the instructions [Installing PyPlot in a robust way](https://github.com/ModiaSim/ModiaMath.jl/wiki/Installing-PyPlot-in-a-robust-way).
 
-Modia3D visualizes the movement of 3D objects with a renderer.
-Currently, the (free) community or the (commercial) professional version of the
-[DLR Visualization](http://www.systemcontrolinnovationlab.de/the-dlr-visualization-library/) library
-are supported. To install the free version for *Windows* or for *Linux* perform the following steps:
+Modia3D visualizes the movement of 3D objects with a renderer. Currently, the (free) community or the (commercial) professional version of the [DLR Visualization](http://www.systemcontrolinnovationlab.de/the-dlr-visualization-library/) library are supported. To install the free version for *Windows* or for *Linux* perform the following steps:
 
-1. Go to [https://visualization.ltx.de/](https://visualization.ltx.de/),
-   provide your contact information and click on *Request download* for *Community Edition*.
-   Afterwards, you get a link to download the library and you need to unzip the file.
+1. Go to [https://visualization.ltx.de/](https://visualization.ltx.de/), provide your contact information and click on *Request download* for *Community Edition*. Afterwards, you get a link to download the library and you need to unzip the file.
 
-2. In your Julia **startup file** (`HOME/.julia/config/startup.jl`) include the environment variable
-   `ENV["DLR_VISUALIZATION"] = "<path-to-library>/Visualization/Extras/SimVis"`.
-   Make sure that the SimVis executable under this directory has execution rights.
-   For example in Linux with command:
-   `chmod +x <path-to-library>/Visualization/Extras/SimVis`
+2. In your Julia **startup file** (`HOME/.julia/config/startup.jl`) include the environment variable `ENV["DLR_VISUALIZATION"] = "<path-to-library>/Visualization/Extras/SimVis"`. Make sure that the SimVis executable under this directory has execution rights. For example in Linux with command:`chmod +x <path-to-library>/Visualization/Extras/SimVis`
 
-3. Start Julia and run one of the examples, for example
-   `include("$(Modia3D.path)/examples/dynamics/Simulate_DoublePendulumWithDampers.jl")`
+3. Start Julia and run one of the examples, for example   `include("$(Modia3D.path)/examples/dynamics/Simulate_DoublePendulumWithDampers.jl")`
 
-If Modia3D cannot use one of the renderers above, it will continue with renderer **NoRenderer**
-where result animation is switched off.
+If Modia3D cannot use one of the renderers above, it will continue with renderer **NoRenderer** where result animation is switched off.
 
 ## Usage of Modia3D by Example
[…]
