### 2019-09-26 c0fc8e2
contactMaterial.jl deleted, changes in readme
