+++ src/contactDetection/ContactDetectionMPR/ContactDetectionMPR_handler.jl
@@ -12,7 +12,6 @@ export updateNoContactPair!
 
 using DataStructures
 
-#const Dict1ValueType = Tuple{Union{SVector{3,Float64},NOTHING}, Union{SVector{3,Float64},NOTHING}, Union{SVector{3,Float64},NOTHING}, Union{Object3D,NOTHING}, Union{Object3D,NOTHING}, Union{Float64,NOTHING} }
 
 const PairID = Int64
 
