+++ test/collision/Collision_Bars.jl
@@ -1,160 +1,0 @@
-module Collision_Bars
-
-using Modia3D
-
-import Modia3D.ModiaMath
-using  Modia3D.StaticArrays
-using  Modia3D.Unitful
-
-
-
-vgreen  = Modia3D.Material(color="Green", transparency=0.5)
-vyellow = Modia3D.Material(color="Yellow", transparency=0.5)
-vred    = Modia3D.Material(color="Red", transparency=0.5)
-cmat    = "Steel"
-
-
-@signal negMove() begin
-   output = ModiaMath.RealScalar("output", causality=ModiaMath.Output, numericType=ModiaMath.WR)
-end
-function Modia3D.computeSignal(signal::negMove, sim::ModiaMath.SimulationState)
-    signal.output.value  = -Modia3D.linearMovement(pi/3, sim.tStart, sim.tEnd, sim.time)
-end
-
-@signal posMove() begin
-   output = ModiaMath.RealScalar("output", causality=ModiaMath.Output, numericType=ModiaMath.WR)
-end
-function Modia3D.computeSignal(signal::posMove, sim::ModiaMath.SimulationState)
-   signal.output.value  = Modia3D.linearMovement(pi/3, sim.tStart, sim.tEnd, sim.time)
-end
-
-
-# Materials
-massProperties = "Aluminium"
-vmat1   = Modia3D.Material(color="LightBlue", transparency=0.5, wireframe=false)
-vblue   = Modia3D.Material(color="LightBlue" , transparency=0.5)
-vred    = Modia3D.Material(color="Red", transparency=0.5)
-vgreen  = Modia3D.Material(color="Green", transparency=0.5)
-vblack  = Modia3D.Material(color="Black", transparency=0.5)
-vyellow = Modia3D.Material(color="Yellow", transparency=0.5)
-vpink   = Modia3D.Material(color="Pink", transparency=0.5)
-vgrey = Modia3D.Material(color="Grey", transparency=0.5)
-cmat    = "Steel"
-vMatAABB = Modia3D.Material(color="Grey", transparency=0.95)
-vrevolute    = Modia3D.Material(color="Red", transparency=0.0)
-
-# dimensions
-Dx = 0.2
-Lx = 1.0
-Ly = 0.3
-Lz = 0.3
-
-groundWidth  = 1.0
-groundHeight = 0.01
-cyl  = Modia3D.Cylinder(1.2*Ly, Ly/2)
-Tcyl = ModiaMath.rot1(90u"°")
-TGround = ModiaMath.rot1(-90u"°")
-
-
-
-# Bounding Boxes AABB
-@assembly BoundingBoxes(world) begin
-  blau1AABB      = Modia3D.Object3D(world,  Modia3D.Box(0.9,0.9,0.9      , material=vMatAABB))
-  blau2AABB      = Modia3D.Object3D(world,  Modia3D.Box(0.9,0.9,0.9      , material=vMatAABB))
-  gruen1AABB     = Modia3D.Object3D(world,  Modia3D.Box(0.9,0.9,0.9      , material=vMatAABB))
-  gruen2AABB     = Modia3D.Object3D(world,  Modia3D.Box(0.9,0.9,0.9      , material=vMatAABB))
-end
-
-@assembly greenAssembly(world) begin
[…]
