+++ test/runexamples.jl
@@ -17,6 +17,7 @@ include(joinpath(dynamicsPath, "Simulate_FallingBall1.jl"))
 #include(joinpath(dynamicsPath, "Simulate_FallingBall4.jl"))
 #include(joinpath(dynamicsPath, "Simulate_FallingBodies.jl"))
 include(joinpath(dynamicsPath, "Simulate_Pendulum.jl"))
+include(joinpath(dynamicsPath, "Simulate_Pendulum_explained.jl"))
 include(joinpath(dynamicsPath, "Simulate_Pendulum_withoutInertia.jl"))
 include(joinpath(dynamicsPath, "Simulate_PendulumWithController.jl"))
 include(joinpath(dynamicsPath, "Simulate_PendulumWithDamper.jl"))

[…]
