### 2019-09-27 f53cfa9
README file and documentation improved
