+++ README.md
@@ -1,4 +1,4 @@
-# Modia3D.jl
+# Modia3D
 
 [![Travis](https://travis-ci.org/ModiaSim/Modia3D.jl.svg?branch=master)](https://travis-ci.org/ModiaSim/Modia3D.jl)
 [![AppVoyer](https://ci.appveyor.com/api/projects/status/github/ModiaSim/Modia3D.jl?svg=true)](https://ci.appveyor.com/project/MartinOtter/modia3d-jl)
@@ -8,25 +8,32 @@
 [![The MIT License](https://img.shields.io/badge/license-MIT-brightgreen.svg?style=flat-square)](https://github.com/ModiaSim/Modia3D.jl/blob/master/LICENSE.md)
 
 
-## Modia Project
-The [Modelica standard library](https://github.com/modelica/ModelicaStandardLibrary) supports the modeling of 3-dimensional multi-body systems with its sub library Modelica.Mechanics.MultiBody. There have been several attempts to improve this library with regards to visualization, collision handling or support of larger models. Over the years it was recognized that the technology of current [Modelica](https://www.modelica.org) has some natural limitations.
+## Modia Platform
 
-Therefore, the open source [Modia](https://github.com/ModiaSim/Modia.jl) project was launched as a domain-specific extension of the  [Julia](https://julialang.org) programming language. The equation-based modeling language is called Modia as well as the experimental [modeling environment](https://github.com/ModiaSim) consisting of several Julia packages. The intention is to utilize the results of this research and prototyping in the design of the next Modelica language generation.
+The *Modia platform* is a prototype system for next generation modeling and simulation of physical systems described by differential and algebraic equations. It consists currently of the following Julia packages that are all under development (not all are yet publicly available):
+
+- Modia - Equation based modeling
+- Modiator - 2D/3D web-app model editor
+- ModiaMath - Simulation enviroment
+- Modia3D - 3D geometry and 3D mechanics
+- ModiaMedia - Thermodynamic property models
+- Modelia - Modelica model importer
 
 
 ## Modia3D
-One part of the Modia project is [Modia3D](https://github.com/ModiaSim/Modia3D.jl). Initially, it is an experimental modeling and simulation environment for 3D mechanical systems, but it shall be expanded into other domains in the future.
 
-Ideas from modern computer game engines are used to achieve a highly flexible setup of mechanical systems including collision handling. Other features are utilized from multi-body programs, such as hierarchical structuring, support for closed kinematic loops, and elastic response calculation.
-The underlying mathematical formulation are hybrid Differential Algebraic Equations (DAEs) that are solved with the variable-step solver IDA via the [Sundials.jl](https://github.com/JuliaDiffEq/Sundials.jl) Julia package.
+Modia3D provides 3D geometry to physical systems so that geometrical objects can be
+directly accessed and utilized in a model. Functions are provided, for example, to compute the volume,
+mass, and inertia of a geometrical object or the distance between two objects.
+Furthermore, Modia3D models 3D mechanical systems and shall be expanded into other domains in the future
+(for example to utilize the 3D geometry to model heat flow in buildings or satellites).
+In the future it will be possible, for example, to model the 3D mechanical part of a robot with Modia3D and the electrical motors and gearboxes driving the joints with Modia.
 
[…]
