### 2019-09-27 38e480f
changes in readme and solidMaterial.jl
