+++ src/Solids/solidMaterial.jl
@@ -8,123 +8,11 @@
 #
 #    Material properties of solids
 
-"""
-    material1 = SolidMaterial(;kwargs...)
-    material2 = SolidMaterial(name)
-
-`SolidMaterial(;kwargs...)` generates a SolidMaterial object by providing the material properties of a solid with
-keyword arguments. Arguments that are not provided have value = NaN.
-
-`SolidMaterial(name)` returns a SolidMaterial object from dictionary Modia3D.solidMaterialPalette using
-the name as dictionary key.
-
-# Arguments
-- `name::AbstractString`: Name of the material (used as key in dictionary Modia3D.solidMaterialPalette)
-
-# Keyword Arguments
-- `density::Float64` in [kg/m^3]: Density, see [Wikipedia](https://en.wikipedia.org/wiki/Density).
-- 'YoungsModulus::Float64` in [Pa]: Youngs's modulus, see [Wikipedia](https://en.wikipedia.org/wiki/Young%27s_modulus).
-- `PoissonsRatio::Float64`: Poisson's ratio, see [Wikipedia](https://en.wikipedia.org/wiki/Poisson%27s_ratio).
-- `meltingPoint::Float64` in [K]: Melting point, see [Wikipedia](https://en.wikipedia.org/wiki/Melting_point).
-                                  If the material is destroyed before its melting point (e.g. wood that is burning)
-                                  then `meltingPoint` is the temperature when destruction of the solid starts.
-- `specificHeatCapacity::Float64` in [J/(kg.K)]: Specific heat capacity, see [Wikipedia](https://en.wikipedia.org/wiki/Heat_capacity).
-- `thermalConductivity::Float64` in [W/(m.K)]: Thermal conductivity, see [Wikipedia](https://en.wikipedia.org/wiki/Thermal_conductivity) and
-                                [List of thermal conductivities](https://en.wikipedia.org/wiki/List_of_thermal_conductivities)
-- `linearThermalExpansionCoefficient::Float64` in [1/K]: Linear thermal expansion coefficient, see [Wikipedia](https://en.wikipedia.org/wiki/Thermal_expansion).
-- `coefficientOfRestitution::Float64`: Coefficient of restitution, see [Wikipedia](https://en.wikipedia.org/wiki/Coefficient_of_restitution).
-- `slidingFrictionCoefficient::Float64`: Kinetic/sliding friction coefficient, see [Wikipedia](https://en.wikipedia.org/wiki/Friction)
-
-# Example
-```julia
-import Modia3D
-mat1 = Modia3D.SolidMaterial(density=3000.0, YoungsModulus=2e11)
-mat2 = Modia3D.SolidMaterial("Steel")    # mat2.density = 8000.0
-```
-"""
-struct SolidMaterial
-   density::Float64              # [kg/m^3]  , https://en.wikipedia.org/wiki/Density
-   YoungsModulus::Float64        # [Pa]      , https://en.wikipedia.org/wiki/Young%27s_modulus
-   PoissonsRatio::Float64        # []        , https://en.wikipedia.org/wiki/Poisson%27s_ratio
-   meltingPoint::Float64         # [K]       , https://en.wikipedia.org/wiki/Melting_point
-                                 # If the material is destroyed before its melting point (e.g. wood that is burning)
-                                 # then meltingPoint is the temperature when destruction starts.
-   heatCapacity::Float64         # [J/(kg.K)], https://en.wikipedia.org/wiki/Heat_capacity
[…]
