### 2019-09-27 d70dfe3
Documentation and README.md file improved/corrected
