+++ docs/src/index.md
@@ -1,29 +1,40 @@
 # Modia3D
 
-[Modia3D](https://github.com/ModiaSim/Modia3D.jl) is a Julia package to model fixed and
-moving objects in 3D (*e.g.* visual shapes, rigid bodies).
-These objects are driven kinematically by pre-defined time functions or are moving dynamically by
-solving Differential Algebraic Equations (DAEs)
-with a variable-step DAE solver.
+## Modia Platform
 
-Collision handling with elastic response calculation is
-performed for objects that are defined with a contact material and (a) have a convex geometry, or
-(b) can be approximated by a set of convex geometries, or (c) have a concave geometry
-that is (automatically) approximated by its convex hull.
-A more detailed overview of the available features is given in the
-[Modia3D documentation](https://ModiaSim.github.io/Modia3D.jl/stable).
-Papers about Modia3D:
+The *Modia platform* is a prototype system for the next modeling and simulation generation of physical systems described by differential and algebraic equations. It consists currently of the following Julia packages that are all under development (not all are yet publicly available):
 
-- *[Collision Handling with Variable-Step Integrators](../resources/documentation/CollisionHandling_Neumayr_Otter_2017.pdf)* ([EOOLT 2017, December](http://www.eoolt.org/2017/))
-- *[Component-Based 3D Modeling of Dynamic Systems](http://www.ep.liu.se/ecp/154/019/ecp18154019.pdf)*([American Modelica Conference 2018, October](https://www.modelica.org/events/modelica2018Americas/index_html))
-- *[Algorithms for Component-Based 3D Modeling](http://www.ep.liu.se/ecp/157/039/ecp19157039.pdf)*([13th International Modelica Conference 2019, March](https://modelica.org/events/modelica2019))
+- *[Modia](https://github.com/ModiaSim/Modia.jl)* - Equation based modeling
+- *Modiator* - 2D/3D web-app model editor
+- *[ModiaMath](https://github.com/ModiaSim/ModiaMath.jl)* - Simulation enviroment
+- *[Modia3D](https://github.com/ModiaSim/Modia3D.jl)* - 3D geometry and 3D mechanics
+- *[ModiaMedia](https://github.com/ModiaSim/ModiaMedia.jl)* - Thermodynamic property models
+- *Modelia* - [Modelica model](https://www.modelica.org/modelicalanguage) importer
 
 
-Before releasing version 1.0, Modia3D shall be
-easily combinable with [Modia](https://github.com/ModiaSim/Modia.jl), for example to define a controlled
-electrical motor with Modia, and add 3D behavior/visualization with Modia3D.
-By this approach the best of both worlds can be combined:
-Special 3D algorithms (Modia3D) + power/flexibility of equation based modeling (Modia).
+## Modia3D
+
+Modia3D provides 3D geometry to physical systems so that geometrical objects can be
+directly accessed and utilized in a model. Functions are provided, for example, to compute the volume,
+mass, and inertia of a geometrical object or the distance between two objects.
+Furthermore, Modia3D models 3D mechanical systems and shall be expanded into other domains in the future
+(for example to utilize the 3D geometry to model heat flow in buildings or satellites).
+In the future it will be possible, for example, to model the 3D mechanical part of a robot with Modia3D and the electrical motors and gearboxes are driving the joints with Modia.
+
[…]
