### 2019-09-27 ad16ddd
Screenshot of Youbot robot added to README.md file
