+++ README.md
@@ -127,9 +127,14 @@ The following code shows how to execute some of these examples.
   include("$(Modia3D.path)/examples/dynamics/Simulate_Pendulum_explained.jl")
   include("$(Modia3D.path)/examples/dynamics/Simulate_DoublePendulumWithDampers.jl")
   include("$(Modia3D.path)/examples/collisions/Simulate_NewtonsCradle.jl")
+  include("$(Modia3D.path)/examples/collisions/Simulate_YouBot.jl")
   include("$(Modia3D.path)/examples/kinematics/Move_FourBar.jl")
   include("$(Modia3D.path)/examples/visual/Move_AllVisualObjects.jl")
 ```
+
+The next figure shows a screenshot of the animation of example `Simulate_YouBot.jl `:
+
+![PendulumPlot](docs/resources/images/youbot-robot.jpg)
 
 
 ## Run tests
