### 2019-09-27 9ff443a
Minor improvement of Youbot robot example
