### 2019-09-27 a0b2a02
isless operator overloaded for contactPairMaterialPalette in order that sorted listing of the palette is possible
