### 2019-09-27 9104140
Minor corrections in Documentation
