### 2019-09-27 1d151f9
Minor correction of README and Documentation
