+++ README.md
@@ -27,7 +27,7 @@ directly accessed and utilized in a model. Functions are provided, for example, 
 mass, and inertia of a geometrical object or the distance between two objects.
 Furthermore, Modia3D models 3D mechanical systems and shall be expanded into other domains in the future
 (for example to utilize the 3D geometry to model heat flow in buildings or satellites).
-In the future it will be possible, for example, to model the 3D mechanical part of a robot with Modia3D and the electrical motors and gearboxes are driving the joints with Modia.
+In the future it will be possible, for example, to model the 3D mechanical part of a robot with Modia3D and the electrical motors and gearboxes that are driving the joints with Modia.
 
 Modia3D uses ideas from modern computer game engines to achieve a highly flexible setup of mechanical systems including collision handling. Other features are utilized from multi-body programs, such as support for closed kinematic loops, and elastic response calculation. The underlying mathematical formulation are hybrid Differential Algebraic Equations (DAEs) that are solved with the variable-step solver IDA via the [Sundials.jl](https://github.com/JuliaDiffEq/Sundials.jl) Julia package.
 
@@ -144,7 +144,7 @@ The next figure shows a screenshot of the animation of example `Simulate_YouBot.
   include("$(Modia3D.path)/test/runtests.jl")
 ```
 
-Additionally to unit tests defined in `runtests.jl`, there are further test examples available.
+Additionally to unit tests defined in `runtests.jl`, there are further test examples available
 under `$(Modia3D.path)/test`.
 
 
