+++ latest/search_index.js
@@ -1,451 +1,3 @@
-var documenterSearchIndex = {"docs": [
-
-{
-    "location": "index.html#",
-    "page": "Home",
-    "title": "Home",
-    "category": "page",
-    "text": ""
-},
-
-{
-    "location": "index.html#Modia3D-1",
-    "page": "Home",
-    "title": "Modia3D",
-    "category": "section",
-    "text": "Modia3D is a Julia package to model fixed and moving objects in 3D (e.g. visual shapes, rigid bodies). These objects are driven kinematically by pre-defined time functions or are moving dynamically by solving Differential Algebraic Equations (DAEs) with a variable-step DAE solver.Collision handling with elastic response calculation is performed for objects that are defined with a contact material and (a) have a convex geometry, or (b) can be approximated by a set of convex geometries, or (c) have a concave geometry that is (automatically) approximated by its convex hull. A more detailed overview of the available features is given in the Modia3D documentation. Papers about Modia3D:Collision Handling with Variable-Step Integrators (EOOLT 2017, December)\nComponent-Based 3D Modeling of Dynamic Systems(American Modelica Conference 2018, October)\nAlgorithms for Component-Based 3D Modeling(13th International Modelica Conference 2019, March)Before releasing version 1.0, Modia3D shall be easily combinable with Modia, for example to define a controlled electrical motor with Modia, and add 3D behavior/visualization with Modia3D. By this approach the best of both worlds can be combined: Special 3D algorithms (Modia3D) + power/flexibility of equation based modeling (Modia)."
-},
-
-{
-    "location": "index.html#Package-Features-1",
-    "page": "Home",
-    "title": "Package Features",
-    "category": "section",
-    "text": "A 3D object is an instance of struct Modia3D.Object3D and defines a coordinate system moving in 3D together with associated data and properties. The following Object3Ds are currently supported:"
-},
-
-{
-    "location": "index.html#Object3Ds-with-a-solid-part-1",
-    "page": "Home",
-    "title": "Object3Ds with a solid part",
-    "category": "section",
-    "text": "Solid parts can be associated with a Modia3D.Object3D. They are defined with struct Modia3D.Solid consisting of an optional solid geometry:(Image: Solids)and other optional properties:mass propreties (defined by geometry+material-name, geometry+density, or directly defined mass properties),\ncontact material (for elastic response calculation),\nvisualization material (for visualization, see below).Since the solid geometry itself is optional, it is possible to just define a coordinate system with associated mass and inertia matrix.The following functions are provided for a solid geometry geo that is associated with an Object3D object3D:volume(geo),\ncentroid(geo),\ninertiaMatrix(geo, mass),\nboundingBox(geo, <other arguments>),\nsupportPoint(geo, <other arguments>),\nisVisible(object3D, renderer),\nhasMass(object3D),\ncanCollide(object3D),\nand other functions."
-},
-
-{
-    "location": "index.html#Object3Ds-for-visualization-1",
-    "page": "Home",
-    "title": "Object3Ds for visualization",
-    "category": "section",
[…]
