### 2019-09-27 4e6c6d8
Fixed issue by using PyPlot explicitly in an example
