### 2019-09-27 e6a74e8
All examples removed from Unit tests that use PyPlot
