### 2019-09-27 675dbed
Removed an example from a unit test that fails on Mac
