### 2019-10-03 #43 @dawbarton
I should also note that I have run the examples on Julia 1.3-RC2 (along with the visualisation software) and everything appears to work appropriately.
