### 2019-10-04 #43 @MartinOtter
> We will soon release a new version of Modia3D (latest end of next week) that will contain a cleaned-up version of the code used as basis for the paper and also a cleaned-up version of the documentation.

This version of Modia3D was released on Sept. 28.
