### 2019-10-04 9a99793
Missing reference added to documentation of damping coefficient
