+++ latest/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"index.html#Modia3D-1","page":"Home","title":"Modia3D","text":"","category":"section"},{"location":"index.html#Modia-Platform-1","page":"Home","title":"Modia Platform","text":"","category":"section"},{"location":"index.html#","page":"Home","title":"Home","text":"The Modia platform is a prototype system for the next modeling and simulation generation of physical systems described by differential and algebraic equations. It consists currently of the following Julia packages that are all under development (not all are yet publicly available):","category":"page"},{"location":"index.html#","page":"Home","title":"Home","text":"Modia - Equation based modeling\nModiator - 2D/3D web-app model editor\nModiaMath - Simulation enviroment\nModia3D - 3D geometry and 3D mechanics\nModiaMedia - Thermodynamic property models\nModelia - Modelica model importer","category":"page"},{"location":"index.html#Modia3D-2","page":"Home","title":"Modia3D","text":"","category":"section"},{"location":"index.html#","page":"Home","title":"Home","text":"Modia3D provides 3D geometry to physical systems so that geometrical objects can be directly accessed and utilized in a model. Functions are provided, for example, to compute the volume, mass, and inertia of a geometrical object or the distance between two objects. Furthermore, Modia3D models 3D mechanical systems and shall be expanded into other domains in the future (for example to utilize the 3D geometry to model heat flow in buildings or satellites). In the future it will be possible, for example, to model the 3D mechanical part of a robot with Modia3D and the electrical motors and gearboxes that are driving the joints with Modia.","category":"page"},{"location":"index.html#","page":"Home","title":"Home","text":"Modia3D uses ideas from modern computer game engines to achieve a highly flexible setup of mechanical systems including collision handling. Other features are utilized from multi-body programs, such as support for closed kinematic loops, and elastic response calculation. The underlying mathematical formulation are hybrid Differential Algebraic Equations (DAEs) that are solved with the variable-step solver IDA via the Sundials.jl Julia package.","category":"page"},{"location":"index.html#","page":"Home","title":"Home","text":"Collision handling with elastic response calculation is performed for geometrical objects that are defined with a contact material and have a convex geometry or are approximated by the convex hull of a concave geometry. Penetration depths and Euclidean distances are computed with the improved Minkowski Portal Refinement (MPR) algorithm [1]. The details of the contact law are provided in [5] and several examples are discussed in [4]. The user's view of Modia3D is introduced in [2] showing the very flexible definition of 3D systems. Some key algorithms are discussed in [3]. A more detailed overview of the available features is also given in the Modia3D documentation.","category":"page"},{"location":"index.html#","page":"Home","title":"Home","text":"Papers and videos about Modia3D:","category":"page"},{"location":"index.html#","page":"Home","title":"Home","text":"[1] Collision Handling with Variable-Step Integrators (EOOLT 2017, December)\n[2] Component-Based 3D Modeling of Dynamic Systems (American Modelica Conference 2018, October)\n[3] Algorithms for Component-Based 3D Modeling (13th International Modelica Conference 2019, March)\n[4] Modia3D: Modeling and Simulation of 3D-Systems in Julia (JuliaCon 2019, July, under review); talk recorded on YouTube\n[5] Collision Handling with Elastic Response Calculation and Zero-Crossing Functions (accepted for publication at (EOOLT 2019, November)","category":"page"},{"location":"index.html#Package-Features-1","page":"Home","title":"Package Features","text":"","category":"section"},{"location":"index.html#","page":"Home","title":"Home","text":"A 3D object is an instance of struct Modia3D.Object3D and defines a co
[…]
