### 2019-10-04 #43 @matbesancon
thanks @dawbarton and @Datseris for the reviews.
@MartinOtter please make the appropriate changes to the paper, we should be finished soon with the review here
