### 2019-10-04 c062919
Documentation improved (and wrong links corrected)
