+++ src/Composition/scene.jl
@@ -127,6 +127,11 @@ include(joinpath(Modia3D.path, "src", "contactDetection", "ContactDetectionMPR",
 
 #-------------------------------------- Gravity field ----------------------------------
 
+"""
+    gravityField = NoGravityField()
+
+Generate an instance of type `NoGravityField` that defines no gravity.
+"""
 struct NoGravityField <: Modia3D.AbstractGravityField
    gvec::SVector{3,Float64} # [m/s^2] Vector of gravity acceleration
    NoGravityField() = new(SVector{3,Float64}(0.0, 0.0, 0.0))
@@ -169,7 +174,7 @@ const EarthRadius = 6.3781e6     # [m]           Radius of earth (https://en.wik
 
 
 """
-    PointGravityField(mass), PointGravityField(;mue=G*EarthMass)
+    PointGravityField([mass|;mue=G*EarthMass])
 
 Return a PointGravityField struct with the gravity field constant mue (mue = G*mass).
 
@@ -202,6 +207,86 @@ gravityAcceleration(grav::PointGravityField, r_abs::AbstractVector) = -(grav.mue
 
 #-------------------------------------- Global Scene Options -------------------------------
 
+"""
+    sceneOptions = Modia3D.SceneOptions(;kwargs...)
+
+Define global options for a simulation of the scene with keyword arguments:
+
+| Keyword arguments             | defaults                                |
+|:------------------------------|:----------------------------------------|
+| enableContactDetection        | true                                    |
+| elasticContactReductionFactor | 1.0                                     |
+| gravityField                  | [`Modia3D.UniformGravityField`](@ref)() |
+| enableVisualization           | true                                    |
+| visualizeGravity              | true                                    |
+| visualizeFrames               | false                                   |
+| visualizeConvexHulls          | true                                    |
+| visualizeContactPoints        | false                                   |
+| visualizeSupportPoints        | false                                   |
+| nominalLength                 | 1.0                                     |
+| defaultFrameLength            | 0.2*nominalLength                       |
+| defaultJointLength            | nominalLength/10                        |
+| defaultJointWidth             | nominalLength/20                        |
+| defaultForceLength            | nominalLength/10                        |
+| defaultForceWidth             | nominalLength/20                        |
+| defaultBodyDiameter           | nominalLength/9                         |
+| defaultWidthFraction          | 20                                      |
+| defaultArrowDiameter          | nominalLength/40                        |
+| `defaultN_to_m`               | 1000                                    |
+| `defaultNm_to_m`              | 1000                                    |
+| useOptimizedStructure         | true                                    |
+| defaultContactSphereDiameter  | 0.1                                     |
+| contactDetection              | ContactDetectionMPR_handler()           |
+
+
+# Optional keyword arguments
+
+- `enableContactDetection::Bool`: = true, if contact detection is enabled
+- `elasticContactReductionFactor::Float64`: used_contact_compliance = contact_compliance * elasticContactReductionFactor (> 0).
+- `gravityField::Modia3D.AbstractGravityField`: Type of gravity field, such as:
+  [`Modia3D.NoGravityField`](@ref),
+  [`Modia3D.UniformGravityField`](@ref),
+  [`Modia3D.PointGravityField`](@ref).
+- `enableVisualization::Bool`: = true, if animation is enabled
+- `visualizeGravity::Bool`: = true, if gravity field shall be visualized (acceleration vector or field center)
[…]
