### 2019-10-04 8b92250
Documentation corrected/improved
