### 2019-10-04 #43 @MartinOtter
> A final comment/question, more from scientific curiousity than anything else, is it possible to use other impact laws other than linear elasticity for the contacts? For example the nonlinear relationships provided by Stronge's Impact Mechanics (Cambridge Press, 2018).

In the previous release of Modia3D, contacts had been defined with linear elasticity. In the release from Sept. 28 and in the paper, contact is defined by a new, nonlinear contact law which is an extension/improvement of existing approaches (including Hertz pressure). The details are documented in the largely improved Modia3D documentation ([collision docu](https://modiasim.github.io/Modia3D.jl/latest/man/CollisionHandling.html))
