### 2019-10-14 #43 @matbesancon
yes it's a pity but it's not showing it, some others have reported it. @vchuravy should we modify the bibstyle?
