### 2019-10-22 #43 @matbesancon
bump @AndreaNeumayr @MartinOtter we should finish this quickly now to get in the first issue
