+++ paper/ref.bib
@@ -6,6 +6,7 @@
   number={1},
   pages={65--98},
   year={2017},
+  doi={10.1137/141000671},
   publisher={SIAM}
 }
 
@@ -17,6 +18,7 @@
 	publisher = {LiU Electronic Press},
 	year =	{2009},
 	month = {Sept.},
+	doi = {10.3384/ecp09430056},
 	url = {http://www.ep.liu.se/ecp/043/062/ecp09430056.pdf}
 }
 
@@ -29,6 +31,7 @@
 	publisher = {LiU Electronic Press},
 	year =	{2014},
 	month = {March},
+	doi = {10.3384/ECP14096899},
 	url = {http://www.ep.liu.se/ecp/096/094/ecp14096094.pdf}
 }
 
@@ -41,6 +44,7 @@
 	publisher = {Springer},
 	doi = { 10.1007/978-3-319-47169-3_15},
 	year =	{2016},
+	doi = {10.1007/978-3-319-47169-3_15},
 	month = {Oct.}
 }
 
@@ -52,6 +56,7 @@
 	publisher = {LiU Electronic Press},
 	year =	{2017},
 	month = {May},
+	doi = {10.3384/ecp17132693},
 	url = {http://www.ep.liu.se/ecp/132/076/ecp17132693.pdf},
 }
 
@@ -65,6 +70,7 @@
 	pages = {9--18},
 	numpages = {10},
 	publisher = {ACM},
+	doi = {10.1145/3158191.3158193},
 	url = {https://modiasim.github.io/Modia3D.jl/resources/documentation/CollisionHandling_Neumayr_Otter_2017.pdf},
 }
 
@@ -75,6 +81,7 @@
 	editor =	{M. Tiller and H. Tummescheit and L. Vanfretti},
 	year = {2018},
 	month = {Oct.},
+	doi = {10.3384/ECP18154175},
 	url = {https://elib.dlr.de/124126/1/2018_Modelica_Modia3D.pdf},
 }
 
@@ -84,6 +91,7 @@
 	booktitle = {Proceedings of the 13th International Modelica Conference},
 	publisher = {LiU Electronic Press},
 	year = {2019},
+	doi = {10.3384/ecp19157383},
 	month = {March},
 }
 
@@ -95,6 +103,7 @@
 	number={3},
 	pages={357--375},
 	year={2011},
+	doi={10.1007/s11044-010-9237-4},
 	publisher={Springer}
 }
 
@@ -105,6 +114,7 @@
 	volume={145},
 	pages={171--187},
 	year={2018},
+	doi={10.1016/j.ijmecsci.2018.07.010},
 	publisher={Elsevier}
 }
 
@@ -125,6 +135,7 @@
 	number = {2},
 	pages = {193-203},
 	year = {1988},
+	doi = {10.1109/56.2083},
 	URL = {https://graphics.stanford.edu/courses/cs448b-00-winter/papers/gilbert.pdf},
 }
 
@@ -174,6 +185,7 @@
   volume =	31,
   number =	3,
   pages =	{363--396},
+  doi = {10.1145/1089014.1089020},
   month =	sep,
   review =	{20100928},
 }
@@ -186,7 +198,7 @@
     volume={5},
     number={1},
 	publisher = {Ubiquity Press},
-	doi = { 10.5334/jors.151},
+	doi = {10.5334/jors.151},
 	year =	{2017},
 }
 
@@ -209,6 +221,6 @@
 	publisher = {LiU Electronic Press},
 	year =	{2015},
 	month = {Sept.},
-   doi = {10.3384/ecp15118427},
+    doi = {10.3384/ecp15118427},
 	url = {http://www.ep.liu.se/ecp/article.asp?issue=118&article=46}
 }
