### 2019-10-24 #43 @MartinOtter
> bump @AndreaNeumayr @MartinOtter we should finish this quickly now to get in the first issue

I am sorry, we had been very busy. We wil update the paper today or latest tomorrow (DOIs are already updated in the ref.bib file).
