### 2019-10-24 be0edd0
Paper improved based on reviewers comments
