+++ paper/header.tex
@@ -6,5 +6,6 @@
 \author[1]{Martin Otter}
 \affil[1]{DLR, Institute of System Dynamics and Control, Germany}
 
-\keywords{Julia, Modia, Modia3D, Modelica, collision handling, Minkowski Portal Refinement algorithm, component-based modeling, elastic force law, response calculation}
+\keywords{Julia, Modia, Modia3D, Modelica, collision handling,
+component-based modeling, elastic response calculation}
 
