+++ paper/sections/3_collisionHandling.tex
@@ -10,20 +10,8 @@ collisions of \emph{many} objects are supported in \emph{real-time} and that a s
 reasonably realistic. Game engines typically make compromises regarding physics (see for
 example \cite{Erez2015}, where it is pointed out that NVIDIA PhysX and Havok neglect
 Coriolis forces).
-When designing an industrial product, it is important that a model of the collision between objects
-is \emph{validated by measurements}. Therefore, the simulation results must
+When designing an industrial product, it is important that collision models can be \emph{validated by measurements}. Therefore, the simulation results must
 be much closer to reality as in a game.
-
-In Modia3D, a currently used algorithm computes the closest distance and/or the penetration depth
-between two \emph{convex} objects.
-If two objects are penetrating each other, two contact forces $\vec{f}_n, \vec{f}_t$ and 
-a contact torque $\vec{\tau}_{\omega}$ are computed from the penetration depth and the relative movement
-and are applied at one point of the contact area (see Fig.~\ref{fig:forces}).
-The advantages of this approach compared to a response calculation with impulses
-is that (a) simulation results are closer
-to reality and (b) it is easier to treat complex contact situations correctly. For example, if several objects have contact at the same time instant, as in the billiard table benchmark of section~\ref{sec:Examples}, 
-it is difficult to solve a model with an impulsive response calculation. The drawback of an elastic
-response calculation is that simulation is often much slower.
 
 
 
@@ -53,7 +41,8 @@ This model is defined with the following Julia program
 (\texttt{r} is the position vector from the parent -- defined with the first
 argument -- to the reference frame of the object and \texttt{fixed} defines
 whether the object is fixed or not fixed in space):
-%
+\newpage
+
 \begin{lstlisting}[language = Julia]
 @assembly BouncingBall begin
   world = Object3D()
@@ -71,18 +60,9 @@ model  = SimulationModel(bounce)
 result = simulate!(model; stopTime=3.0)
 \end{lstlisting}
 
-\begin{figure}[h]
-	\centering
-	\includegraphics[width=0.25\textwidth]{figures/forces.pdf}
-	\caption{Contact normal force $\vec{f}_n$, contact tangential force $\vec{f}_t$ 
-		(= sliding friction force) and contact torque $\vec{\tau}_{\omega}$
-		between two penetrating objects. $\vec{e}_n, \vec{e}_t, \vec{e}_{\omega}$ are unit vectors
-		in direction of the respective relative movement.}
-	\label{fig:forces}
-\end{figure}
 
 %
-\begin{figure}[htb]
+\begin{figure}[tb]
 	\centering
 	\includegraphics[width=0.2\textwidth]{figures/bouncingBall.png}
 	\caption{Ball falling on a table.}
@@ -169,6 +149,16 @@ distribution in tangential direction and that the response characteristics provi
 approximation of the resultant force and torque of these distributions.
 %
 
+\begin{figure}[t]
+	\centering
+	\includegraphics[width=0.25\textwidth]{figures/forces.pdf}
+	\caption{Contact normal force $\vec{f}_n$, contact tangential force $\vec{f}_t$ 
+		(= sliding friction force) and contact torque $\vec{\tau}_{\omega}$
+		between two penetrating objects. $\vec{e}_n, \vec{e}_t, \vec{e}_{\omega}$ are unit vectors
+		in direction of the respective relative movement.}
+	\label{fig:forces}
+\end{figure}
+
 %
 The MPR-algorithm computes the contact point, $\delta$, and a unit vector $\vec{e}_n$ that is orthogonal
 to the contacting surfaces. The novel response characteristic of (\ref{eq:response})
[…]
