### 2019-10-24 #43 @AndreaNeumayr
@matbesancon We've improved our paper and updated it based on the reviewers comments. We've also improved the documentation of our package and the readme in our latest release of Modia3D v0.4.0.
I'm sorry, it took us a bit longer than expected.
