### 2019-10-24 #43 @matbesancon
Ok thanks @dawbarton @Datseris would you like to have another round of review and see if the last criteria & boxes are checked?
