### 2019-10-24 #43 @MartinOtter
**The Paper was improved based on reviewers comments.**

>    In the paper and documentation there is a lack of comparison with other software that achieve similar characteristics. The software should be put into context of other existing software, and also attempt to educate the potential user on why they should choose this software, instead of other. I find it hard to believe that Modia3D.jl is the only software that does what it does.

This is now described better in the introduction to explain the differences to other software and also giving links to the large number of Game engines and multibody programs. This description could still be improved, but then other parts of the paper have to be removed due to the limit of 6 pages.

> In the start of paragraph 2 the authors "compare" Modia3D.jl with game engines, but this is not a true comparison. They only mention that they use game developing paradigms without going into much depth. Me, as a potential user, should I consider comparing your software with Unreal Engine or with DifferentialEquations.jl ?

It was tried to make this more clear. Again, due to limited space the description is short. Modia3D is not intented to be a game engine (but use some good ideas from them).

>    I think the authors assume knowledge of what Modia or Modelica is. The reader of the paper may not necessarily have this knowledge (I for example do not). I think the authors should try to describe Modia in a couple of sentences before using it in text.

We added a table of the Modia platform with a short description and added some examples for the intended applications.

>    Both the paper as well as the documentation lack a statement of need. This also ties with my first comment; the authors do not really put the work into context.

Added some explanations in the introduction.

>    There is no discussion about performance but I think the authors should consider providing one. Does this software match the performance of its competitors? Does it exceed it?

As stated in the conclusion at the end, we put effort in functionality and did not yet tune it for performance.

> Furthermore, novel algorithms seem to have been developed and be in use for this software (the three papers associated with the software, e.g. "Collision Handling with Variable-Step Integrators (EOOLT 2017, December)". Do these approaches also offer performance benefits?

>    The documentation is severely lacking examples as well as instructions. The package features are listed in detail, but instructions of how to leverage them are nowhere to be seen. Specifically, the "getting started page" is simply installation instructions, while all the "examples" of the software are presented like so:

    # Examples to demonstrate kinematic movement of assemblies
    include("$(Modia3D.path)/examples/kinematics/Move_DoublePendulum.jl")
    include("$(Modia3D.path)/examples/kinematics/Move_FourBar.jl")
    include("$(Modia3D.path)/examples/kinematics/Move_FourBar2.jl")

    This is not enough. At least one example must be expanded in detail and every single step of the example must be explained, also in detail. In fact, at the moment I cannot see how a user would actually use Modia3D.jl, given the absence of guidance. Please be aware that listing an example is not enough: you rely on me being able to deduce what a command does by seeing its output. This is not the proper way to learn, as I might deduce many things wrongly.

The README.md file of Modia3D and the documentation has been significantly improved (version 0.4.0 of Modia3D)

>    Although optional, I think the authors should strongly consider bundling this plotting library they use with their software. The current situation is a no-go for me.

Probably, you mean 3D animation (not plotting)? In the paper the description was added that another team is developing an open source 2D/3D web app (Modiater).

> Similar to Datseris' review my main criticism of the paper is the lack of context. What other rigid body dynamics packages should this be compared with? (Open source or closed source - a few reference points would be helpful.) While game engine comparisons are interesting (particularly what can be learnt from the structure of the code), I doubt they are the most appropriate in terms of the scientific objectives. Would software packages such as Mambo (http://danko.mechanical.illinois.edu/Mambo/MamboHome.htm) be an appropriate comparison? (I don't have any experience of it myself.)

Yes, we agree, the description could be better. We improved the introduction. Note, there are hundreds of multibody programs (smaller like Mambo and large ones like SIMPACK). However, none of them models non-mechanical parts directly (such as electrical motors, or an air-conditioning system). We tried to explain shortly in the introduction of the paper, that Modia3D will be an integrated part in a larger context.

> Further to the lack of context, there could be a bit more information about the intended use cases of the software. From my own experience I can infer what they are likely to be but making it a bit more explicit would be good.

Added some examples in the introduction.

> There are also a number of unqualified assertions in the paper, for example "The advantages of this approach compared to a response calculation with impulses is that (a) simulation results are closer to reality and (b) it is easier to treat complex contact situations correctly."; references supporting these statements would be much appreciated.

Added some more explanations here.

> The online documentation provided is a little sparse, but well above what I would consider to be the bare minimum. With the information provided, I'd be happy to have a go at creating my own models though I suspect I would end up digging through the source code to find out necessary details.

> The remaining item I haven't checked off is the references: the guidelines above appear to indicate that DOIs are required for all relevant publications and as far as I can see there aren't any provided. Other than that, the references look appropriate.

Added DOIs, but the latex templates do not use them.

> A final comment/question, more from scientific curiousity than anything else, is it possible to use other impact laws other than linear elasticity for the contacts? For example the nonlinear relationships provided by Stronge's Impact Mechanics (Cambridge Press, 2018).

The documentation was now significantly enhanced (Modia3D 0.4.0) and has nonlinear elastic contact response calculation
