### 2019-10-24 #43 @Datseris
Hello,

I've went through the new version of the paper which I consider vastly improved; I accept for publication.
