### 2019-11-04 #43 @matbesancon
ping @dawbarton for the last checks in the list
