### 2019-11-05 #43 @matbesancon
Some quick things:
Table 1: Julia packages of **the** Modia platform.
Page one col 2, "Whereas" is a bit awkward at the beginning of a sentence, prefer a comma with the previous sentence.
> On the other hand, multi-body programs are usually far from the
flexible setup of games.

Why? The sentence would benefit from a bit more justification and context
