### 2019-11-05 #43 @matbesancon
Page 3 col 2, replace cheap and cheaply with "low-cost", low computational cost or similar
