### 2019-11-05 #43 @matbesancon
Page 4 equation (3)
```
julia> 2.125 - 70 * 0.03
0.02499999999999991
```

Would be worth it to mention the order of magnitude of  the resisual, 0.03m*s^-1 (rounded up) can be considered non-zero in some contexts
