### 2019-11-06 3c969ae
Some minor textual improvements performed as suggested by @whedon
