+++ paper/sections/2_flexibleDefinition.tex
@@ -2,6 +2,7 @@
 % Section 
 %%%%%%%%%%%%%%%%%%%%%%%
 \section{Flexible Definition of 3D Systems}
+\label{sect_flexibleDefinition}
 
 Modia3D follows the approach of modern game engines to provide
 a coordinate system as a primitive that is located in 3D and has a 
