### 2019-11-06 a82c263
Removed unused comments
