### 2019-11-06 #43 @MartinOtter
>
>
> Some quick things:
> Table 1: Julia packages of **the** Modia platform.
> Page one col 2, "Whereas" is a bit awkward at the beginning of a sentence, prefer a comma with the previous sentence.

Fixed

>
> > On the other hand, multi-body programs are usually far from the
> > flexible setup of games.
>
> Why? The sentence would benefit from a bit more justification and context

Added a reference to section 2. Due to lack of space, did not add more.

For information: Multi-body programs usually have "body" objects and "joints" and then associate attributes with a body (e.g. "markers" on the body or "visual" information representing the body).
Games have "coordinate systems" as main objects and associate attributes with a coordinate system,
such as geometry, visualization, dynamics, collision properties, light, camera, sound, video, etc.).
The latter allows a much more flexible definition.
