### 2019-11-06 #43 @MartinOtter
>
>
> Page 3 col 2, replace cheap and cheaply with "low-cost", low computational cost or similar
Fixed.
