### 2019-11-06 #43 @MartinOtter
>
>
> Page 4 equation (3)
>
> ```
> julia> 2.125 - 70 * 0.03
> 0.02499999999999991
> ```
>
> Would be worth it to mention the order of magnitude of the resisual, 0.03m*s^-1 (rounded up) can be considered non-zero in some contexts

Made no change due to lack of space. The numbers of these equations have been just visually extracted from the plot window and it is just a sketch to explain the bend in the figure.
The actual numbers with more digits give an exact zero.
