### 2019-11-06 #43 @MartinOtter
>
>
> [ point_right Check article proof page_facing_up point_left ](https://github.com/JuliaCon/proceedings-papers/blob/jcon.00043/jcon.00043/10.21105.jcon.00043.pdf)

Looks fine (just DOIs still missing in the literature section)
