### 2019-11-11 0724cb6
Modia3D preparing for combination with Modia (SimulationModel splited in model and simulationState)
