+++ src/Composition/dynamics.jl
@@ -58,6 +58,103 @@ function get_WorldObject3D(assembly::Modia3D.AbstractAssembly)::Object3D
 end
 
 
+
+"""
+    model = Modia3D.Model(assembly::Modia3D.AbstractAssembly;
+                          analysis::ModiaMath.AnalysisType=ModiaMath.DynamicAnalysis)
+
+Generate a `Modia3D.Model` from an `assembly` generated with macro [`Modia3D.@assembly`](@ref)
+and the type of `analysis` to be carried out on the `assembly`.
+"""
+struct Model
+    name::String
+    nz::Int   # Number of event indicators
+    assembly::Modia3D.AbstractAssembly
+    var::ModiaMath.ModelVariables
+    analysis::ModiaMath.AnalysisType
+    x_start::Vector{Float64}
+    x_fixed::Vector{Bool}
+    x_nominal::Vector{Float64}
+    is_constraint::Vector{Bool}
+
+    function Model(assembly::Modia3D.AbstractAssembly;
+                   analysis::ModiaMath.AnalysisType=ModiaMath.DynamicAnalysis)
+
+        name = Modia3D.trailingPartOfName( string( typeof(assembly) ) )
+        world = get_WorldObject3D(assembly)
+        assembly._internal.referenceObject3D = world
+        # println("\n... world reference frame: ", ModiaMath.fullName(world))
+
+        # Determine Variables in assembly
+        var = ModiaMath.ModelVariables(assembly, analysis=analysis)
+        #ModiaMath.print_ModelVariables(var)
+
+        # Set initial values for x
+        x_start   = zeros(var.nx)
+        x_fixed   = fill(false,var.nx)
+        x_nominal = fill(1.0  ,var.nx)
+        ModiaMath.copy_start_to_x!(var, x_start, x_fixed, x_nominal)
+        # println("... x0 = ", x_start)
+
+        # Last nfc equations are the constraint equations
+        is_constraint = fill(false, var.nx)
+        for i = (var.nx-var.nfc+1):var.nx
+            is_constraint[i] = true
+        end
+
+        # Construct Scene(..) object
+        so = assembly._internal.sceneOptions
+        sceneOptions::SceneOptions = typeof(so) == NOTHING ? SceneOptions() : so
+        scene = Scene(sceneOptions)
+        scene.analysis = analysis
+        assembly._internal.scene = scene
+        if !scene.options.enableContactDetection
+            scene.collide = false
+        end
+
+        # Build tree for optimized structure or standard structure
+        # collision handling is only available for optimized structure
+        nz = 0
+        if scene.options.useOptimizedStructure
+            build_superObjs!(scene, world)
+            if !scene.options.enableContactDetection
+               scene.collide = false
+            end
+
+            if scene.collide
+                initializeContactDetection!(world, scene)
+                if scene.collide
+                    nz = scene.options.contactDetection.contactPairs.nz
+                    append!(scene.allVisuElements, world.contactVisuObj1)
+                    append!(scene.allVisuElements, world.contactVisuObj2)
+                    append!(scene.allVisuElements, world.supportVisuObj1A)
+                    append!(scene.allVisuElements, world.supportVisuObj1B)
+                    append!(scene.allVisuElements, world.supportVisuObj1C)
+                    append!(scene.allVisuElements, world.supportVisuObj2A)
+                    append!(scene.allVisuElements, world.supportVisuObj2B)
+                    append!(scene.allVisuElements, world.supportVisuObj2C)
+                end
+            end
+            initializeMassComputation!(scene)
+        else
+            build_tree!(scene, world)
+            if scene.options.enableContactDetection
[…]
