### 2019-11-12 dff5c30
Preparing for combining Modia with Modia3D (Modia3D: model_f1, model_f2 functions provided)
