+++ src/Composition/dynamics.jl
@@ -58,10 +58,126 @@ function get_WorldObject3D(assembly::Modia3D.AbstractAssembly)::Object3D
 end
 
 
+"""
+    realScalar = getFlangeVariable(assembly::Modia3D.AbstractAssembly, name::String)
+
+Return the reference of a Revolute or Prismatic flange variable, given the top-level `assembly`
+and the full path `name` of the flange variable. If one of the following flange variable names
+is provided, a reference to the respective RealScalar is returned:
+```
+flange_b.phi
+flange_b.s
+flange_b.tau
+flange_b.f
+der(<full-name>.flange_b.phi)
+der(<full-name>.flange_b.s)
+```
+
+If one of the following flange names is provided, `nothing` is returned:
+```
+flange_a.phi
+flange_a.s
+flange_a.tau
+flange_a.f
+der(<full-name>.flange_a.phi)
+der(<full-name>.flange_a.s)
+```
+"""
+function getFlangeVariable(assembly::Modia3D.AbstractAssembly, name::String)::Union{ModiaMath.RealScalar,Nothing}
+    # Handle der(..)
+    lenName = length(name)
+    if lenName >= 4 && SubString(name,1,4) == "der("
+        if SubString(name,lenName,lenName) != ")" || lenName < 6
+            error("Wrong name = \"", name, "\"")
+        end
+        name2 = SubString(name,5,lenName-1)
+        der   = true
+    else
+        name2 = name
+        der   = false
+    end
+
+    field = assembly
+    while true
+        #println("... before. name2 = ", name2)
+        i = findnext(".", name2, 1)
+        #println("... after.")
+        if i==nothing || length(name2) <= i[1] || i[1] <= 1
+            error("Name not found: \"", name, "\"")
+        end
+        fieldName = Symbol(SubString(name2,1,i[1]-1))
+        name2     = SubString(name2,i[1]+1)
+        if !isdefined(field, fieldName)
+            error("Name not found: \"", name, "\"")
+        end
+        field = getfield(field, fieldName)
+
+        # Handle the names "flange_a" and "flange_b"
+        #println("... field name = ", fieldName, ", typeof(field) = ", typeof(field))
+        if name2=="flange_a.phi"
+            if typeof(field) == Modia3D.Composition.TreeJointRevolute
+                return nothing
+            else
+                error("Name = \"", name, "\" is no flange variable of a Revolute joint")
+            end
+        elseif name2=="flange_a.tau"
+            if typeof(field) == Modia3D.Composition.TreeJointRevolute
+                return nothing
+            else
+                error("Name = \"", name, "\" is no flange variable of a Revolute joint")
+            end
+        elseif name2=="flange_a.s"
+            if typeof(field) == Modia3D.Composition.TreeJointPrismatic
+                return nothing
+            else
+                error("Name = \"", name, "\" is no flange variable of a Prismatic joint")
+            end
+        elseif name2=="flange_a.f"
+            if typeof(field) == Modia3D.Composition.TreeJointPrismatic
+                return nothing
+            else
+                error("Name = \"", name, "\" is no flange variable of a Prismatic joint")
+            end
+        elseif name2=="flange_b.phi"
+            if typeof(field) == Modia3D.Composition.TreeJointRevolute
+                return der ? getfield(field,:w) : getfield(field,:phi)
+            else
+                error("Name = \"", name, "\" is no flange variable of a Revolute joint")
+            end
+        elseif name2=="flange_b.tau"
[…]
