### 2019-11-12 b0e1c3b
Version number in Project.toml file of master adapter (0.4.1-dev)
