### 2020-06-09 #43 @AndreaNeumayr
>
>
> @matbesancon
> I added all available DOI's for the reference list (ref.bib) in the paper. The JuliaCon proceedings template is not supporting it. There are no DOI's shown. Do you know how to fix that?

We discussed about the DOI's already.

Thank you, now worries.
