### 2020-08-10 #43 @AndreaNeumayr
@vchuravy there's no archival DOI. But is it possible to tag it with the release number of Modia3D? Release v0.4.0 https://github.com/ModiaSim/Modia3D.jl/releases/tag/v0.4.0
