### 2020-08-10 #43 @ChrisRackauckas
Quickly setup a Zenodo?
