### 2020-08-10 #43 @matbesancon
Yes you will need to setup a zenodo and archive the current version, github itself is not an archiving process
