### 2020-08-11 #43 @AndreaNeumayr
Thanks, Modia3D v0.4.0 has a DOI now: 10.5281/zenodo.3978914
