### 2020-08-12 #43 @AndreaNeumayr
> [ 👉 Check article proof 📄 👈 ](https://raw.githubusercontent.com/JuliaCon/proceedings-papers/jcon.00043/jcon.00043/10.21105.jcon.00043.pdf)

this looks fine. There are no problems with labels, references, and citations
