### 2020-08-12 6c80a59
Update paper.yml
