+++ paper/paper.yml
@@ -18,4 +18,4 @@ affiliations:
   - name: DLR, Institute of System Dynamics and Control, Germany
     index: 1
 date: 7 July 2019
-bibliography: ref
+bibliography: ref.bib
