### 2020-08-12 #43 @arfon
@MartinOtter - I think this PR should fix the issue here: https://github.com/ModiaSim/Modia3D.jl/pull/13
